% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_prevalence.R
\name{adjusted_prevalence}
\alias{adjusted_prevalence}
\title{Adjust Prevalence in a Microbiome Object}
\usage{
adjusted_prevalence(
  obj,
  method = "min",
  output_file = "adjusted_prevalence.rds"
)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object.}

\item{method}{A character string specifying the method ("min", "mean", "median", "max").}

\item{output_file}{A character string specifying the file path to save the adjusted object.}
}
\value{
An adjusted object with pruned taxa.
}
\description{
Adjusts the prevalence of ASVs based on a specified method
and prunes those that do not meet the prevalence threshold.
}
\examples{
\donttest{
if (requireNamespace("DspikeIn", quietly = TRUE)) {
  data("physeq_16SOTU", package = "DspikeIn")

  # Adjust prevalence using the "min" method
  adjusted_physeq <- adjusted_prevalence(physeq_16SOTU, method = "min")
  print(adjusted_physeq)

  # Convert phyloseq to TreeSummarizedExperiment (TSE)
  tse_16SOTU <- convert_phyloseq_to_tse(physeq_16SOTU)

  # Adjust prevalence using the "max" method on a TSE object
  adjusted_tse <- adjusted_prevalence(tse_16SOTU, method = "max")
  print(adjusted_tse)
}
}
}
