% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spike_percentage_hashcodes.R
\name{calculate_spike_percentage_hashcodes}
\alias{calculate_spike_percentage_hashcodes}
\title{Calculate Spike Percentage for Hashcodes}
\usage{
calculate_spike_percentage_hashcodes(
  physeq,
  spiked_species,
  identifier_type = "species",
  output_path = NULL,
  passed_range = c(0.1, 11)
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbiome data.}

\item{spiked_species}{A character vector of spiked species or hashcodes to be analyzed.}

\item{identifier_type}{A character string specifying whether the identifiers are "species" or "hashcode". Default is "species".}

\item{output_path}{A character string specifying the output file path for the Word document. Default is NULL.}

\item{passed_range}{A numeric vector of length 2 specifying the range of percentages that are considered "passed". Default is c(0.1, 11).}
}
\value{
A data frame containing the merged data with the calculated percentages and pass/fail results.
}
\description{
This function calculates the percentage of reads for specified spiked species or hashcodes relative to the total reads in each sample. 
It merges the specified taxa, calculates the read percentages, and categorizes the results based on a specified range.
The results are saved as a flextable in a Word document and as a CSV file.
Note: Calculating the spiked species percentage needs to be done after merging all ASVs/OTUs rooting from spiked species.
}
\examples{
# Calculate spike percentage for specified hashcodes
merged_data <- calculate_spike_percentage_hashcodes(spiked_16S, hashcodes, output_path = NULL, passed_range = c(0.1, 11))
}
