% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spike_percentage_species.R
\name{calculate_spike_percentage_species}
\alias{calculate_spike_percentage_species}
\title{Calculate Spike Percentage for Specified Species in Phyloseq Object}
\usage{
calculate_spike_percentage_species(
  physeq,
  spiked_species,
  identifier_type = "species",
  output_path = NULL,
  passed_range = c(0.1, 11)
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbial data.}

\item{spiked_species}{A character vector of spiked species to check in the phyloseq object.}

\item{identifier_type}{A character string specifying the type of identifier to use ("species" or "hashcode"). Default is "species".}

\item{output_path}{A character string specifying the path to save the output files. Default is NULL, which saves the report as "merged_data.docx".}

\item{passed_range}{A numeric vector of length 2 specifying the range of percentages to categorize results as "passed". Default is c(0.1, 11).}
}
\value{
A data frame containing the percentage of spiked taxa reads and the pass/fail results.
}
\description{
This function calculates the percentage of reads from specified spiked species in a phyloseq object.
It merges the spiked taxa into one ASV, calculates the percentage of reads, categorizes the results as passed or failed,
and saves the results as a DOCX and CSV file.
}
\examples{
# Example usage:
# Define the spiked species
spiked_species <- c("Tetragenococcus_halophilus", "Tetragenococcus_sp.")

# Calculate spike percentage and generate the report
calculate_spike_percentage_species(spiked_16S, spiked_species, identifier_type = "species", output_path = NULL, passed_range = c(0.1, 11))

# Save the results
merged_data <- calculate_spike_percentage_species(spiked_16S, spiked_species)
}
