% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_differential_abundance.R,
%   R/normalization_set.R
\name{convert_categorical_to_factors}
\alias{convert_categorical_to_factors}
\title{Convert Categorical Columns to Factors in Sample Data}
\usage{
convert_categorical_to_factors(ps)

convert_categorical_to_factors(ps)
}
\arguments{
\item{ps}{A phy#'}
}
\value{
A phyloseq object with updated sample data.

A phyloseq object with updated sample data.
}
\description{
This function converts categorical columns in sample data to factors.
}
\examples{
\dontrun{
ps <- phyloseq::GlobalPatterns
ps_factor <- convert_categorical_to_factors(ps)
}loseq object.
}
