% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tse_to_phyloseq.R
\name{convert_tse_to_phyloseq}
\alias{convert_tse_to_phyloseq}
\title{Convert a \code{TreeSummarizedExperiment} to a \code{phyloseq} Object}
\usage{
convert_tse_to_phyloseq(tse)
}
\arguments{
\item{tse}{A \code{TreeSummarizedExperiment} object containing:
\itemize{
\item \strong{OTU Table (\code{counts})}.
\item \strong{Taxonomy Table (\code{rowData})}.
\item \strong{Sample Metadata (\code{colData})}.
\item \strong{Phylogenetic Tree (\code{rowTree}, optional)}.
\item \strong{Reference Sequences (\code{metadata}, optional)}.
}}
}
\value{
A \code{phyloseq} object containing all preserved components.
}
\description{
Converts a \code{TreeSummarizedExperiment} (TSE) object back
into a \code{phyloseq} object, ensuring the phylogenetic tree and
reference sequences are restored, if present.
}
\examples{
if (requireNamespace("DspikeIn", quietly = TRUE)) {
  data("physeq_16SOTU", package = "DspikeIn")

  # Convert phyloseq to TSE
  tse_16SOTU <- convert_phyloseq_to_tse(physeq_16SOTU)

  # Convert TSE back to phyloseq
  phy_M <- convert_tse_to_phyloseq(tse_16SOTU)
  print(phy_M)
}
}
