% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package Startup Message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{The library name (automatically passed by R).}

\item{pkgname}{The package name (automatically passed by R).}
}
\description{
Displays a message when the DspikeIn package is loaded.
}
\keyword{internal}
