% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-ExpData_class.R
\docType{class}
\name{ExpData-class}
\alias{ExpData-class}
\title{ExpData Virtual Class}
\description{
This is a virtual superclass to support S4 object structure in DESeqDataSet extension.
}
\keyword{internal}
