% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long_format_data.R
\name{get_long_format_data}
\alias{get_long_format_data}
\title{Convert a Phyloseq or TSE Object into a Long-Format Data Frame}
\usage{
get_long_format_data(obj)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object.}
}
\value{
A long-format \code{data.frame} containing taxonomic, abundance, and sample metadata.
}
\description{
Converts a \code{phyloseq} or \code{TreeSummarizedExperiment} object into
a long-format data frame, similar to \code{phyloseq::psmelt()}, for compatibility
with visualization functions such as \code{alluvial_plot()}.
}
\examples{
if (requireNamespace("DspikeIn", quietly = TRUE)) {
  data("physeq_16SOTU", package = "DspikeIn")
  tse_16SOTU <- convert_phyloseq_to_tse(physeq_16SOTU)

  # Turn it to the long format
  melted_ph <- get_long_format_data(physeq_16SOTU)
  melted <- get_long_format_data(tse_16SOTU)
}
}
