% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_phy_tree}
\alias{get_phy_tree}
\title{Extract Phylogenetic Tree}
\usage{
get_phy_tree(obj)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object.}
}
\value{
A phylogenetic tree object.
}
\description{
Retrieves the phylogenetic tree.
}
