% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_adj.R
\name{get_sample_sums}
\alias{get_sample_sums}
\title{Extract Sample Sums from Object}
\usage{
get_sample_sums(obj)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object.}
}
\value{
A numeric vector of total counts per sample.
}
\description{
Retrieves the total sequence counts per sample from a \code{phyloseq} or \code{TreeSummarizedExperiment} object.
}
