% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_graphml.R
\name{load_graphml}
\alias{load_graphml}
\title{Load GraphML Without 'id' Conflicts}
\usage{
load_graphml(filename = "herp.spiecsym.network.graphml")
}
\arguments{
\item{filename}{Name of the GraphML file (default: "herp.spiecsym.network.graphml") or full file path.}
}
\value{
An \code{igraph} object representing the network.
}
\description{
Loads a GraphML network from \code{DspikeIn} or user path, ensuring all attributes remain intact and correctly assigning original node names.
}
