% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{data}
\name{methods}
\alias{methods}
\title{Available Normalization Methods}
\format{
A character vector of method names.
}
\usage{
data(methods)
}
\description{
A dataset containing the normalization methods implemented in DspikeIn.
}
\examples{
data(methods)
print(methods)
}
\keyword{datasets}
