% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_colors.R
\docType{data}
\name{MG_colors}
\alias{MG_colors}
\title{Predefined Color Palette for Consistent Plot Styling}
\format{
An object of class \code{character} of length 84.
}
\usage{
MG_colors
}
\value{
A character vector of hex color codes.
}
\description{
A color vector containing a predefined set of colors
}
\examples{
\dontrun{
# Use MG_colors in ggplot2
  ggplot2::ggplot(mtcars, aes(x = wt, y = mpg, color = factor(cyl))) +
  geom_point(size = 3) +
  scale_color_manual(values = MG_colors[1:3]) +
  my_custom_theme()
}
}
\keyword{datasets}
