% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.DESeq}
\alias{norm.DESeq}
\title{DESeq Normalization with Pseudocount and Integer Conversion}
\usage{
norm.DESeq(obj, groups, pseudocount = 1)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object.}

\item{groups}{A string specifying the grouping variable in sample data.}

\item{pseudocount}{A numeric value added to avoid zeros in the dataset.}
}
\value{
A list containing the normalized object (same format as input) and scaling factors.
}
\description{
DESeq Normalization
}
\examples{
\donttest{
if (requireNamespace("DspikeIn", quietly = TRUE)) {
  data("physeq_16SOTU", package = "DspikeIn")
  result_DESeq <- norm.DESeq(physeq_16SOTU, groups = "Animal.type", pseudocount = 1)
}
}
}
