% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.PoissonSeq}
\alias{norm.PoissonSeq}
\title{PoissonSeq Normalization}
\usage{
norm.PoissonSeq(raw)
}
\arguments{
\item{raw}{A matrix of raw counts.}
}
\value{
A list containing the normalized data and scaling factor.
}
\description{
A normalization method specifically designed for RNA-seq data, available in the PoissonSeq package.
}
\references{
PoissonSeq Package.
}
