% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.RUVg}
\alias{norm.RUVg}
\title{Remove Unwanted Variation Using Control Genes (RUVg)}
\usage{
norm.RUVg(ps, groups)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{groups}{The name of the sample variable in the sample data that indicates group labels.}
}
\value{
A list containing the normalized phyloseq object and adjustment factors.
}
\description{
Remove Unwanted Variation Using Control Genes (RUVg)
}
\references{
Gagnon-Bartsch, Jacob, and Speed 2013; Risso et al. 2014; Gagnon-Bartsch and Speed 2012.
}
