% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.RUVs}
\alias{norm.RUVs}
\title{Remove Unwanted Variation Using Replicate Samples (RUVs)}
\usage{
norm.RUVs(raw, groups)
}
\arguments{
\item{raw}{A matrix of raw counts.}

\item{groups}{A vector of group labels.}
}
\value{
A list containing the normalized data and adjustment factor.
}
\description{
Uses replicate samples to remove unwanted variation.
}
\references{
Gagnon-Bartsch, Jacob, and Speed 2013; Risso et al. 2014; Gagnon-Bartsch and Speed 2012.
}
