% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.SVA}
\alias{norm.SVA}
\title{Surrogate Variable Analysis for Sequencing Data (SVA)}
\usage{
norm.SVA(ps, groups)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{groups}{A vector of group labels.}
}
\value{
A list containing the normalized data and adjustment factor.
}
\description{
Surrogate Variable Analysis for Sequencing Data (SVA)
}
\references{
SVA package.
}
