% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.TMM}
\alias{norm.TMM}
\title{TMM Normalization (Trimmed Mean of M component)}
\usage{
norm.TMM(ps, groups)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{groups}{A string specifying the grouping variable in sample data.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
TMM Normalization (Trimmed Mean of M component)
}
\examples{
\dontrun{
ps <- phyloseq::GlobalPatterns
result_TMM <- norm.TMM(ps, "SampleType")
normalized_ps_TMM <- result_TMM$dat.normed
scaling_factors_TMM <- result_TMM$scaling.factor
}

}
