% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.UQ}
\alias{norm.UQ}
\title{UQ Normalization (Upper Quartile)}
\usage{
norm.UQ(obj, groups)
}
\arguments{
\item{obj}{A phyloseq object.}

\item{groups}{A string specifying the grouping variable in sample data.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
UQ Normalization (Upper Quartile)
}
