% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_phyloseq_clr.R
\name{normalize_phyloseq_clr}
\alias{normalize_phyloseq_clr}
\title{Centered Log-Ratio (CLR)}
\usage{
normalize_phyloseq_clr(
  ps,
  feature_category = c("all", "iqlr", "zero", "lvha"),
  min_counts = 1
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{Character string specifying which features to use as the divisor for the geometric mean calculation.}

\item{min_counts}{Minimum number of counts required for a sample to be retained.}
}
\value{
A phyloseq object with CLR normalization.
}
\description{
This function performs CLR normalization on a phyloseq object using log-ratio transformation.
It handles zero, negative, and NA values by removing them and informing the user of the modifications.
}
\examples{
data(GlobalPatterns)
ps_clr <- normalize_phyloseq_clr(GlobalPatterns, feature_category = "iqlr", min_counts = 1000)
}
