% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_phyloseq_cpm.R
\name{normalize_phyloseq_cpm}
\alias{normalize_phyloseq_cpm}
\title{Counts Per Million (CPM)}
\usage{
normalize_phyloseq_cpm(
  ps,
  feature_category = c("all", "iqlr", "zero", "lvha"),
  min_counts = 1
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{Character string specifying which features to use as the divisor for the geometric mean calculation.}

\item{min_counts}{Minimum number of counts required for a sample to be retained.}
}
\value{
A phyloseq object with CPM normalization.
}
\description{
This function normalizes counts in a phyloseq object to counts per million (CPM).
It handles zero, negative, and NA values by removing them and informing the user of the modifications.
}
\examples{
data(GlobalPatterns)
ps_cpm <- normalize_phyloseq_cpm(GlobalPatterns, feature_category = "iqlr", min_counts = 1000)
}
