% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_phyloseq_css.R
\name{normalize_phyloseq_css}
\alias{normalize_phyloseq_css}
\title{Cumulative Sum Scaling (CSS)}
\usage{
normalize_phyloseq_css(
  ps,
  feature_category = c("all", "iqlr", "zero", "lvha"),
  min_counts = 1
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{Character string specifying which features to use as the divisor for the geometric mean calculation.}

\item{min_counts}{Minimum number of counts required for a sample to be retained.}
}
\value{
A phyloseq object with CSS normalization.
}
\description{
This function performs CSS normalization on a phyloseq object based on the cumulative sum of gene abundances.
It handles zero, negative, and NA values by removing them and informing the user of the modifications.
}
\examples{
data(GlobalPatterns)
ps_css <- normalize_phyloseq_css(GlobalPatterns, feature_category = "iqlr", min_counts = 1000)
}
