% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_phyloseq_rarefy.R
\name{normalize_phyloseq_rarefy}
\alias{normalize_phyloseq_rarefy}
\title{Rarefaction: Random Subsampling to the Smallest Library Size}
\usage{
normalize_phyloseq_rarefy(
  ps,
  feature_category = c("all", "iqlr", "zero", "lvha"),
  min_counts = 1
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{Character string specifying which features to use as the divisor for the geometric mean calculation.}

\item{min_counts}{Minimum number of counts required for a sample to be retained.}
}
\value{
A rarefied phyloseq object.
}
\description{
This function performs rarefaction on a phyloseq object by randomly subsampling counts to the smallest library size.
It handles zero, negative, and NA values by removing them and informing the user of the modifications.
}
\examples{
data(GlobalPatterns)
ps_rarefied <- normalize_phyloseq_rarefy(GlobalPatterns, feature_category = "iqlr", min_counts = 1000)
}
