% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_phyloseq_rle.R
\name{normalize_phyloseq_rle}
\alias{normalize_phyloseq_rle}
\title{Relative Log Expression (RLE)}
\usage{
normalize_phyloseq_rle(
  ps,
  feature_category = c("all", "iqlr", "zero", "lvha"),
  min_counts = 1
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{Character string specifying which features to use as the divisor for the geometric mean calculation.}

\item{min_counts}{Minimum number of counts required for a sample to be retained.}
}
\value{
A phyloseq object with RLE normalization.
}
\description{
This function performs RLE normalization on a phyloseq object using a pseudo-reference sample.
It handles zero, negative, and NA values by removing them and informing the user of the modifications.
}
\examples{
data(GlobalPatterns)
ps_rle <- normalize_phyloseq_rle(GlobalPatterns, feature_category = "iqlr", min_counts = 1000)
}
