% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_phyloseq_tmm.R
\name{normalize_phyloseq_tmm}
\alias{normalize_phyloseq_tmm}
\title{Trimmed Mean of M-values (TMM)}
\usage{
normalize_phyloseq_tmm(
  ps,
  feature_category = c("all", "iqlr", "zero", "lvha"),
  min_counts = 1
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{Character string specifying which features to use as the divisor for the geometric mean calculation.}

\item{min_counts}{Minimum number of counts required for a sample to be retained.}
}
\value{
A phyloseq object with TMM normalization.
}
\description{
This function performs TMM normalization on a phyloseq object, based on weighted trimmed mean of log fold-changes.
It handles zero, negative, and NA values by removing them and informing the user of the modifications.
}
\examples{
data(GlobalPatterns)
ps_tmm <- normalize_phyloseq_tmm(GlobalPatterns, feature_category = "iqlr", min_counts = 1000)
}
