% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_phyloseq_tss.R
\name{normalize_phyloseq_tss}
\alias{normalize_phyloseq_tss}
\title{Total Sum Scaling (TSS)}
\usage{
normalize_phyloseq_tss(
  ps,
  feature_category = c("all", "iqlr", "zero", "lvha"),
  min_counts = 1
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{Character string specifying which features to use as the divisor for the geometric mean calculation.}

\item{min_counts}{Minimum number of counts required for a sample to be retained.}
}
\value{
A phyloseq object with TSS normalization.
}
\description{
This function normalizes abundances in a phyloseq object by dividing by the sample library size.
It handles zero, negative, and NA values by removing them and informing the user of the modifications.
}
\examples{
data(GlobalPatterns)
ps_tss <- normalize_phyloseq_tss(GlobalPatterns, feature_category = "iqlr", min_counts = 1000)
}
