% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_phyloseq_vst.R
\name{normalize_phyloseq_vst}
\alias{normalize_phyloseq_vst}
\title{Variance Stabilizing Transformation (VST)}
\usage{
normalize_phyloseq_vst(
  physeq,
  design_formula = ~sample_name,
  feature_category = c("all", "iqlr", "zero", "lvha"),
  min_counts = 1,
  output_file = NULL,
  pseudocount = 1
)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{design_formula}{A formula specifying the design for DESeq2.}

\item{feature_category}{Character string specifying which features to use as the divisor for the geometric mean calculation.}

\item{min_counts}{Minimum number of counts required for a sample to be retained.}

\item{output_file}{An optional file path to save the transformed DESeqDataSet object.}

\item{pseudocount}{A numeric value to add to counts before transformation to avoid log(0).}
}
\value{
A phyloseq object with VST normalization.
}
\description{
This function performs VST normalization on a phyloseq object using the DESeq2 package.
It handles zero, negative, and NA values by removing them and informing the user of the modifications.
}
\examples{
data(GlobalPatterns)
ps_vst <- normalize_phyloseq_vst(GlobalPatterns, design_formula = ~ SampleType, feature_category = "iqlr", min_counts = 1000)
}
