% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_differential_abundance.R
\name{perform_and_visualize_DA_edgeR}
\alias{perform_and_visualize_DA_edgeR}
\title{Perform and Visualize Differential Abundance Analysis with edgeR}
\usage{
perform_and_visualize_DA_edgeR(
  ps,
  group_var,
  contrast,
  pseudocount = 1,
  significance_level = 0.05,
  output_csv_path = "DA_edgeR.csv",
  point_size = 3,
  target_glom = "Genus",
  palette = MG()
)
}
\arguments{
\item{ps}{A phyloseq object containing microbiome data.}

\item{group_var}{A string specifying the grouping variable in sample data.}

\item{contrast}{A vector specifying the levels to compare (e.g., c("Acris", "Anaxyrus")).}

\item{pseudocount}{A numeric value for pseudocount addition. Default is 1.}

\item{significance_level}{A numeric value specifying the significance level for filtering significant OTUs. Default is 0.05.}

\item{output_csv_path}{Path to save the edgeR results as a CSV file. Default is "DA_edgeR.csv".}

\item{point_size}{A numeric value specifying the size of points in the volcano plot. Default is 3.}

\item{target_glom}{A string specifying the taxonomic rank to aggregate taxa. Default is "Genus".}

\item{palette}{A character vector of color hex codes for plotting. Default is `MG()`.}
}
\value{
A list containing the edgeR results, a phyloseq object with significant OTUs/ASVs, and a volcano plot.
}
\description{
This function performs edgeR analysis and generates a volcano plot based on microbiome data.
}
\examples{
\dontrun{
results_edgeR <- perform_and_visualize_DA_edgeR(
  ps = ps, 
  group_var = "Host.genus", 
  contrast = c("Acris", "Anaxyrus"), 
  output_csv_path = "DA_edgeR.csv", 
  target_glom = "Genus", 
  significance_level = 0.05
)

# Visualize Volcano Plot
print(results_edgeR$plot)
}

}
