% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_nj.R
\name{plot_tree_nj}
\alias{plot_tree_nj}
\title{Plot a Neighbor-Joining Phylogenetic Tree with Bootstrap Values}
\usage{
plot_tree_nj(fasta_file, output_file = "neighbor_joining_tree.png")
}
\arguments{
\item{fasta_file}{A character string specifying the path to the FASTA file containing DNA sequences.}

\item{output_file}{A character string specifying the output file path for the plot. Default is "neighbor_joining_tree.png".}
}
\value{
NULL. The function saves the plot to the specified output file and displays it.
}
\description{
This function reads DNA sequences from a FASTA file, constructs a Neighbor-Joining phylogenetic tree,
generates bootstrap values, and plots the tree.
}
\examples{
\dontrun{
# Plot the Neighbor-Joining tree from the "tetra.fasta" file
# The output will be saved as "neighbor_joining_tree_with_bootstrap.png"
plot_tree_nj("tetra.fasta", output_file = "neighbor_joining_tree_with_bootstrap.png")
}
}
