% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_sentence.R
\name{print_sentence}
\alias{print_sentence}
\title{Print a Sentence Multiple Times}
\usage{
print_sentence(sentence)
}
\arguments{
\item{sentence}{A character string specifying the sentence to be printed.}
}
\value{
NULL. The function prints the sentence to the console.
}
\description{
This function prints a specified sentence multiple times (default is 10 times).
}
\examples{
# Example usage:
# Define a sentence
sentence <- " ¯\\\\_(ツ)_/¯  ¯\\\\_(ツ)_/¯  ¯\\\\_(ツ)_/¯  ¯\\\\_(ツ)_/¯ 
# ¯\\\\_(ツ)_/¯ ¯\\\\_(ツ)_/¯ ¯\\\\_(ツ)_/¯ ¯\\\\_(ツ)_/¯ ¯\\\\_(ツ)_/¯  ¯\\\\_(ツ)_/¯ "

# Print the sentence multiple times
print_sentence(sentence)
}
