% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{process_data_with_feature_category}
\alias{process_data_with_feature_category}
\title{Process data with various feature categories}
\usage{
process_data_with_feature_category(
  ps,
  feature_category = c("all", "iqlr", "zero", "lvha")
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{A character string specifying the feature category ("all", "iqlr", "zero", "lvha").}
}
\value{
A pruned phyloseq object based on the specified feature category.
}
\description{
Process data with various feature categories
}
