% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vst_analysis.R
\name{run_vst_analysis}
\alias{run_vst_analysis}
\title{Run Variance Stabilizing Transformation (VST) Analysis}
\usage{
run_vst_analysis(
  physeq,
  design_formula = ~sample_name,
  output_file = NULL,
  pseudocount = 1
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbiome data.}

\item{design_formula}{A formula specifying the design of the experiment. Default is ~ sample_name.}

\item{output_file}{A character string specifying the output file name for the transformed DESeqDataSet object. Default is NULL.}

\item{pseudocount}{A numeric value specifying the pseudocount to add to the counts data before transformation. Default is 1.}
}
\value{
A phyloseq object with the VST-transformed counts data.
}
\description{
This function performs a variance stabilizing transformation (VST) on the counts data of a phyloseq object
using DESeq2, and returns a phyloseq object with the transformed data.
}
\examples{
# Perform VST analysis on a phyloseq object
transformed_physeq <- run_vst_analysis(physeq_ITS)
}
