% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{set_nf}
\alias{set_nf}
\title{Set Normalization Factors in the Sample Data of the Phyloseq Object}
\usage{
set_nf(ps, scaling.factor)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{scaling.factor}{A vector of normalization factors.}
}
\value{
A phyloseq object with updated sample data.
}
\description{
Set Normalization Factors in the Sample Data of the Phyloseq Object
}
\examples{
\dontrun{
ps <- phyloseq::GlobalPatterns
factors <- rep(1, phyloseq::nsamples(ps))
ps <- set_nf(ps, factors)
}
}
