% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_phyloseq_ASV_OTUID.R
\name{summ_phyloseq_ASV_OTUID}
\alias{summ_phyloseq_ASV_OTUID}
\title{Summarize Phyloseq Data Based on ASV_ID}
\usage{
summ_phyloseq_ASV_OTUID(physeq)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbiome data.}
}
\value{
A data frame containing summary statistics for each ASV.
}
\description{
This function generates summary statistics (mean, median, standard deviation, standard error, and quantiles)
for each ASV (Amplicon Sequence Variant) in a phyloseq object.
}
\examples{
\dontrun{
# Summarize the phyloseq data based on ASV_ID
summary_stats <- summ_phyloseq_ASV_OTUID(physeq_ITSOTU)

# Print the summary statistics
print(summary_stats)
}
}
