% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_barplot.R
\name{taxa_barplot}
\alias{taxa_barplot}
\title{Create a Taxa Barplot}
\usage{
taxa_barplot(
  physeq,
  target_glom = "Genus",
  custom_tax_names = NULL,
  normalize = TRUE,
  treatment_variable = "Treatment",
  abundance_type = "relative",
  x_angle = 20,
  fill_variable = target_glom,
  facet_variable = "Phylum",
  top_n_taxa = 20,
  palette = MG()
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbiome data.}

\item{target_glom}{A character string specifying the taxonomic rank to plot (e.g., "Genus").}

\item{custom_tax_names}{A character vector specifying custom taxonomic names for the levels. Default is NULL.}

\item{normalize}{A logical indicating whether to normalize the sample counts to relative abundances. Default is TRUE.}

\item{treatment_variable}{A character string specifying the treatment variable to use for the x-axis. Default is "Treatment".}

\item{abundance_type}{A character string specifying whether to plot "relative" or "absolute" abundance. Default is "relative".}

\item{x_angle}{A numeric value specifying the angle of the x-axis text labels. Default is 20.}

\item{fill_variable}{A character string specifying the variable to use for fill in stacking taxa. Default is target_glom.}

\item{facet_variable}{A character string specifying the variable to use for faceting. Default is "Phylum".}

\item{top_n_taxa}{A numeric value specifying the number of top taxa to include in the plot. Default is 20.}

\item{palette}{A character vector of color hex codes to use for the fill colors. Default is `MG()`.}
}
\value{
A list containing the ggplot2 bar plot object (`barplot`) and the pruned phyloseq object (`taxa_data`) with the top taxa.
}
\description{
This function generates a bar plot of the relative or absolute abundances of taxa at a specified taxonomic rank.
The top taxa are selected, and the plot can be customized with various options.
}
\examples{
# Generate a taxa barplot for the Genus rank with absolute abundance
bp_ab <- taxa_barplot(physeq_16SASV, target_glom = "Genus",
  treatment_variable = "Host.species", abundance_type = "absolute", x_angle = 90,
  fill_variable = "Genus", facet_variable = "Diet", top_n_taxa = 20, palette = MG())
print(bp_ab$barplot)

# Generate a taxa barplot for the Genus rank with relative abundance
bp_rel <- taxa_barplot(physeq_16SASV, target_glom = "Genus", 
  treatment_variable = "Host.genus", abundance_type = "relative", x_angle = 90, 
  fill_variable = "Genus", facet_variable = "Diet", top_n_taxa = 20, palette = MG())
print(bp_rel$barplot)

}
