% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_plot.R
\name{transform_plot}
\alias{transform_plot}
\title{Generate and Save Boxplots with Statistical Tests}
\usage{
transform_plot(
  data,
  x_var,
  y_vars,
  methods_var,
  color_palette = DspikeIn::color_palette$cool_MG,
  output_prefix = "plot",
  width = 15,
  height = 13,
  stat_test = "kruskal.test"
)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{x_var}{A character string specifying the column name for the x variable (categorical/factor).}

\item{y_vars}{A character vector specifying the column names for the y variables.}

\item{methods_var}{A character string specifying the column name for the grouping variable.}

\item{color_palette}{A character string for the color palette from \code{DspikeIn::color_palette$cool_MG}.}

\item{output_prefix}{A character string specifying the prefix for the output file names. Default is \code{"plot"}.}

\item{width}{A numeric value specifying the width of the output plot. Default is \code{15}.}

\item{height}{A numeric value specifying the height of the output plot. Default is \code{13}.}

\item{stat_test}{A character string specifying the statistical test to use (\code{"kruskal.test"} or \code{"anova"}). Default is \code{"kruskal.test"}.}
}
\value{
A list containing the ggplot2 boxplot objects and the statistical comparison results.
}
\description{
This function generates boxplots for multiple y variables and performs statistical
tests (Kruskal-Wallis or ANOVA) to compare groups. It also performs pairwise comparisons when
needed and saves the plots in PNG and PDF formats.
}
\examples{
\donttest{
if (requireNamespace("DspikeIn", quietly = TRUE)) {
  data("methods", package = "DspikeIn")

  # Define variables for the y-axis
  y_vars <- c("Spike.percentage", "Total.reads", "Spike.reads")

  # Generate and save boxplots with Kruskal-Wallis test
  plot_output <- transform_plot(
    data = methods,
    x_var = "Methods",
    y_vars = y_vars,
    methods_var = "Methods",  # Ensure this parameter is needed
    color_palette = DspikeIn::color_palette$cool_MG,
    stat_test = "kruskal.test"
  )

  # Print the output plot
  print(plot_output)
}
}
}
