% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_differential_abundance.R
\name{visualize_differential_abundance}
\alias{visualize_differential_abundance}
\title{Visualize Differential Abundance}
\usage{
visualize_differential_abundance(results, group_var, point_size = 3, palette)
}
\arguments{
\item{results}{A data frame containing the differential abundance results.}

\item{group_var}{A string specifying the grouping variable.}

\item{point_size}{A numeric value specifying the size of points in the volcano plot.}

\item{palette}{A character vector of color hex codes for plotting.}
}
\value{
A ggplot2 object representing the volcano plot.
}
\description{
Visualize Differential Abundance
}
