% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_Network.R
\name{weight_Network}
\alias{weight_Network}
\title{Analyze and Visualize a Microbial Network}
\usage{
weight_Network(graph_path = NULL, save_metrics = TRUE)
}
\arguments{
\item{graph_path}{Character. The \strong{GraphML file path}. Default (\code{NULL}) loads \code{"Complete.graphml"} from \strong{DspikeIn}.
Valid internal options: \code{"Complete.graphml"}, \code{"NoHubs.graphml"}, \code{"NoBasid.graphml"}.
If an \strong{external path} is provided, the function loads that instead.}

\item{save_metrics}{Logical. If \code{TRUE}, saves \code{"Global_Network_Metrics.csv"}.}
}
\value{
A \strong{list} containing:
\item{plot}{A \strong{ggplot2} object displaying the network.}
\item{metrics}{A \strong{data.frame} with global network metrics.}
\item{graph}{The \strong{igraph} object with assigned attributes.}
}
\description{
Loads a microbial network from the \strong{DspikeIn} package (\code{Complete.graphml}, \code{NoHubs.graphml}, or \code{NoBasid.graphml})
or from a user-provided \strong{GraphML} file. Computes network metrics, assigns modularity-based node colors (\code{cool_MG} from \strong{DspikeIn}),
and visualizes the network using \code{ggraph}. Optionally saves computed \strong{global network metrics} as a CSV file.
}
\examples{
\donttest{
if (requireNamespace("DspikeIn", quietly = TRUE)) {
  Complete <- load_graphml("Complete.graphml")

  # Run network weighting function on the loaded dataset
  result <- weight_Network()

  # Load a specific GraphML dataset from DspikeIn
  result <- weight_Network(graph_path = "NoHubs.graphml")

  # Load a custom GraphML file from user directory,
  # for external graphml please use **full address**
  print(result$plot)

  # View network metrics
  result$metrics
}
}

}
