% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcceptableRange.R
\docType{data}
\name{AcceptableRange}
\alias{AcceptableRange}
\title{Acceptable Range Data}
\format{
A data frame with range of microbial abundance
}
\source{
Internal package dataset.
}
\usage{
data(AcceptableRange)
}
\description{
A dataset containing acceptable ranges for microbial abundance estimates.
}
\examples{
data(AcceptableRange)
head(AcceptableRange)
}
\keyword{datasets}
