% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_abundance_one_third.R
\name{adjust_abundance_one_third}
\alias{adjust_abundance_one_third}
\title{Adjust Abundance by One-Third}
\usage{
adjust_abundance_one_third(physeq, factor = 3, output_file = NULL)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbiome data.}

\item{factor}{A numeric value specifying the factor by which to divide the abundance data. Default is 3.}

\item{output_file}{A character string specifying the output file name for the adjusted phyloseq object. Default is NULL.}
}
\value{
A phyloseq object with the adjusted abundance data.
}
\description{
This function normalizes the abundance data in a phyloseq object by dividing each value by a specified factor.
The normalization process is based on the volume of DNA processed for sequencing, such as 16S rRNA amplicons.
}
\examples{
\dontrun{
if(interactive()){
  # Adjust the abundance data by dividing each value by 3
  adjusted_physeq <- adjust_abundance_one_third(physeq16S, factor = 3)
}
}
}
