% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_prevelance.R
\name{adjusted_prevalence}
\alias{adjusted_prevalence}
\title{Adjust Prevalence in a Phyloseq Object}
\usage{
adjusted_prevalence(
  spiked_16S,
  method = "min",
  output_file = "adjusted_prevalence_physeq.rds"
)
}
\arguments{
\item{spiked_16S}{A phyloseq object containing the taxonomic and abundance data.}

\item{method}{A character string specifying the method to calculate the number of reads. Options are "min", "mean", "median", and "max". Default is "min".}

\item{output_file}{A character string specifying the file path to save the adjusted phyloseq object. Default is "adjusted_prevalence_physeq.rds".}
}
\value{
A phyloseq object with adjusted prevalence.
}
\description{
This function adjusts the prevalence of ASVs (Amplicon Sequence Variants) in a phyloseq object
based on a specified method to determine the number of reads (min, mean, median, or max).
It then prunes the ASVs that do not meet the prevalence threshold and saves the adjusted phyloseq object.
}
\examples{
# Adjust prevalence based on the minimum number of reads
spiked_16S_min <- adjusted_prevalence(physeq_16S_adj_scaled, method = "min")

# Adjust prevalence based on the mean number of reads
spiked_16S_mean <- adjusted_prevalence(physeq_16S_adj_scaled, method = "mean")

# Adjust prevalence based on the median number of reads
spiked_16S_median <- adjusted_prevalence(physeq_16S_adj_scaled, method = "median")

# Adjust prevalence based on the maximum number of reads
spiked_16S_max <- adjusted_prevalence(physeq_16S_adj_scaled, method = "max")
}
