% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_plot.R
\name{alluvial_plot}
\alias{alluvial_plot}
\title{Generate an Alluvial Plot}
\usage{
alluvial_plot(
  data,
  axes,
  abundance_threshold = 10000,
  fill_variable = "Phylum",
  silent = TRUE,
  abundance_type = "absolute",
  total_reads = NULL
)
}
\arguments{
\item{data}{A data frame containing the microbial data, including abundance and categorical variables.}

\item{axes}{A character vector specifying the axes (columns) to include in the alluvial plot.}

\item{abundance_threshold}{A numeric value specifying the minimum abundance threshold for including samples. Default is 10000.}

\item{fill_variable}{A character string specifying the variable to use for fill in the alluvial plot. Default is "Phylum".}

\item{silent}{A logical indicating whether to suppress messages from `is_alluvia_form`. Default is TRUE.}

\item{abundance_type}{A character string specifying whether to plot "relative" or "absolute" abundance. Default is "absolute".}

\item{total_reads}{A numeric value specifying the total number of reads for the sample when using relative abundance. Default is NULL.}
}
\value{
A ggplot2 object representing the alluvial plot.
}
\description{
This function creates an alluvial plot to visualize the abundance of taxa across multiple categorical axes.
It filters out samples with abundance below the specified threshold and generates the plot using `ggplot2`.
}
\examples{
# Example with provided data format
meli_Abs_WSal <- read.csv("path/to/meli_Abs_WSal.csv")

# Generate alluvial plot
is_alluvia_form(as.data.frame(meli_Abs_WSal), axes = 1:6, silent = TRUE)
alluvial_plot <- alluvial_plot(data = meli_Abs_WSal, axes = c("Diet", "Host.genus", "Ecoregion.III"), abundance_threshold = 10000, fill_variable = "Phylum", silent = TRUE, abundance_type = "absolute")
print(alluvial_plot)
}
