% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_plot.R
\name{alluvial_plot}
\alias{alluvial_plot}
\title{Create an Alluvial Plot}
\usage{
alluvial_plot(
  data,
  axes = c("Host_genus", "Ecoregion_III", "Diet"),
  abundance_threshold = 10000,
  fill_variable = "Phylum",
  silent = TRUE
)
}
\arguments{
\item{data}{A dataframe containing the data to be plotted.}

\item{axes}{A character vector specifying the columns of the data to be used as axes in the alluvial plot. Default is c("Host_genus", "Ecoregion_III", "Diet").}

\item{abundance_threshold}{A numeric value specifying the minimum abundance threshold for samples to be included in the plot. Default is 10000.}

\item{fill_variable}{A character string specifying the variable used for filling the alluvia. Default is "Phylum".}

\item{silent}{A logical value indicating whether to suppress warnings and messages. Default is TRUE.}
}
\value{
A ggplot object representing the alluvial plot.
}
\description{
This function generates an alluvial plot from the given dataset, illustrating the abundance
across specified factors. Samples with abundance below the given threshold are filtered out.
}
\examples{
# `meli` is a preprocessed dataframe as described in the example
alluvial_plot <- alluvial_plot(data = meli, axes = c("Abundance", "Diet", "Host_genus", "Ecoregion_III"), abundance_threshold = 20000, silent = TRUE)
# Display the plot
print(alluvial_plot)
}
