% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bartlett_test.R
\name{Bartlett_test}
\alias{Bartlett_test}
\title{Perform Bartlett's Test for Homogeneity of Variances with Multiple Transformations}
\usage{
Bartlett_test(
  data,
  grouping_var,
  transformations = c("raw", "sqrt", "log", "zscale")
)
}
\arguments{
\item{data}{A data frame containing the data to test.}

\item{grouping_var}{A character string specifying the column name to use for grouping.}

\item{transformations}{A character vector specifying the transformations to apply. Default is c("raw", "sqrt", "log", "zscale").}
}
\value{
NULL. The function prints the Bartlett test results for each variable and transformation.
}
\description{
This function performs Bartlett's test for homogeneity of variances on numeric variables
within a data frame, using specified transformations. It prints the p-value for each variable
and transformation combination.
}
\examples{
# Example usage:
# Assume `methods` is a data frame with a column "Methods" for grouping
# and other numeric columns to test.
# file_path <- system.file("extdata", "methods.rds", package = "DspikeIn") 
methods <- readRDS(file_path) 
print(methods) 
Bartlett_test(methods, "Methods")
}
