% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_list_average_scaling_factors.R
\name{calculate_list_average_scaling_factors}
\alias{calculate_list_average_scaling_factors}
\title{Calculate Average Scaling Factors for Multiple Spiked Species with Equal or Different Spiked Cell Counts}
\usage{
calculate_list_average_scaling_factors(
  physeq,
  spiked_species_list,
  spiked_cells_list,
  merge_method = c("sum", "max")
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbial data.}

\item{spiked_species_list}{A list of character vectors, where each vector contains the spiked species to merge for each scaling factor calculation.}

\item{spiked_cells_list}{A numeric vector specifying the number of spiked cells for each set of spiked species.}

\item{merge_method}{The method to use for merging ASVs: "sum" or "max".}
}
\value{
A numeric vector of averaged scaling factors.
}
\description{
This function calculates scaling factors for multiple spiked species in a phyloseq object,
merges ASVs for each species if necessary, averages the scaling factors, and returns the
averaged scaling factors. Different spiked cell counts can be provided for each set of spiked species.
}
\examples{
# Example usage:
# Assume `physeq` is a phyloseq object and `spiked_species_list` is a list of character vectors.
# spiked_species_list <- list(c("Species1"), c("Species2"), c("Species3"), c("Species4"))
# spiked_cells_list <- c(1000, 1500, 2000, 2500)
# average_scaling_factors <- calculate_average_scaling_factors(physeq, spiked_species_list, spiked_cells_list, "sum")
}
