% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spike_percentage.R
\name{calculate_spike_percentage}
\alias{calculate_spike_percentage}
\title{Calculate Spike Percentage for Specified Taxa in a Phyloseq Object}
\usage{
calculate_spike_percentage(
  physeq,
  merged_spiked_species = NULL,
  merged_spiked_hashcodes = NULL,
  output_path = "merged_data.docx",
  passed_range = c(0.1, 11)
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbial data.}

\item{merged_spiked_species}{A character vector of spiked species to check in the phyloseq object. Default is NULL.}

\item{merged_spiked_hashcodes}{A character vector of spiked hashcodes to check in the phyloseq object. Default is NULL.}

\item{output_path}{A character string specifying the path to save the output files. Default is "merged_data.docx".}

\item{passed_range}{A numeric vector of length 2 specifying the range of percentages to categorize results as "passed". Default is c(0.1, 11).}
}
\value{
A data frame containing the percentage of spiked taxa reads and the pass/fail results.
}
\description{
This function calculates the percentage of reads from specified spiked species or hashcodes in a phyloseq object.
It merges the spiked taxa into one ASV, calculates the percentage of reads, categorizes the results as passed or failed,
and saves the results as a DOCX and CSV file.
}
\examples{
# Load example data
data(physeq_16SASV)
physeq <- physeq_16SASV

# Define the spiked species
merged_spiked_species <- c("Tetragenococcus_halophilus", "Tetragenococcus_sp.")
calculate_spike_percentage(physeq, merged_spiked_species = merged_spiked_species, passed_range = c(0.1, 10))

# Define the spiked hashcodes
merged_spiked_hashcodes <- c("hashcode1", "hashcode2")
calculate_spike_percentage(physeq, merged_spiked_hashcodes = merged_spiked_hashcodes, passed_range = c(0.1, 10))
}
