% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spike_percentage_list.R
\name{calculate_spike_percentage_list}
\alias{calculate_spike_percentage_list}
\title{Calculate Spike Percentage for List of Specified Taxa in a Phyloseq Object}
\usage{
calculate_spike_percentage_list(
  physeq,
  merged_spiked_species = NULL,
  output_path = "merged_data.docx",
  passed_range = c(0.1, 11)
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing microbial data.}

\item{merged_spiked_species}{A character vector of spiked species to check in the phyloseq object.}

\item{output_path}{A character string specifying the path to save the output files. Default is \code{"merged_data.docx"}.}

\item{passed_range}{A numeric vector of length 2 specifying the range of percentages to categorize results as "passed". Default is \code{c(0.1, 11)}.}
}
\value{
A data frame containing the percentage of spiked taxa reads and the pass/fail results.
}
\description{
This function calculates the percentage of reads from specified spiked species in a \code{phyloseq} object.
It merges the spiked taxa into one ASV, calculates the percentage of reads, categorizes the results as passed or failed,
and saves the results as DOCX and CSV files.
}
\examples{
\dontrun{
# Example usage:
spiked_species_list <- c("Pseudomonas aeruginosa", "Escherichia coli", "Clostridium difficile")
result <- calculate_spike_percentage_list(physeq, merged_spiked_species = spiked_species_list, passed_range = c(0.1, 10))
print(result)
}
}
