% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summary_stats_table.R
\name{calculate_summary_stats_table}
\alias{calculate_summary_stats_table}
\title{Calculate Summary Statistics Table}
\usage{
calculate_summary_stats_table(data, output_path = NULL)
}
\arguments{
\item{data}{A data frame containing the data to be summarized.}

\item{output_path}{A character string specifying the output path for the Word document. Default is NULL, which sets the output path to "post_eval_summary.docx".}
}
\value{
A flextable object containing the summary statistics.
}
\description{
This function calculates summary statistics (mean, standard deviation, standard error, quantiles) for numeric columns in a data frame,
creates a flextable for formatted output, and saves the table in a Word document and CSV file.
}
\examples{
# Calculate summary statistics and save the table
summary_table <- calculate_summary_stats_table(initial_stat_ASV, output_path = "summary.docx")
post_eval_summary <- read.csv("post_eval_summary.csv")
}
