% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{color_palette}
\alias{color_palette}
\title{Original and Extended Color Palette Sequence}
\format{
A list of character vectors of color codes.
}
\description{
This object provides an original color palette sequence (MG) and an extended palette using the rainbow function.
The extended palette can be used for plotting in ggplot2 or other plotting systems.
}
\examples{
# Use the extended palette for your plot
# ggplot2 example:
# ggplot(data, aes(x, y, color = group)) +
#   geom_point() +
#   scale_color_manual(values = extended_palette)
}
