% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclusion.R
\name{conclusion}
\alias{conclusion}
\title{Generate Spike Success Report and Summary Statistics}
\usage{
conclusion(
  physeq,
  merged_spiked_species,
  max_passed_range = 11,
  output_path = NULL
)
}
\arguments{
\item{physeq}{A `phyloseq::phyloseq` object containing the microbial data.}

\item{merged_spiked_species}{A character vector of spiked species to check in the phyloseq object.}

\item{max_passed_range}{A numeric value specifying the maximum acceptable spike percentage. Default is 11.}

\item{output_path}{A character string specifying the path to save the output files. Default is NULL, which saves the report as "spike_success_report.docx".}
}
\value{
A `data.frame` containing the summary statistics of the spike success report.
}
\description{
This function calculates the spike success report for a given phyloseq object, 
summarizes the results, and saves the summary statistics in both DOCX and CSV formats.
It uses the `flextable` package to create and format the tables in the DOCX file.
}
\examples{
\dontrun{
# Example usage:
merged_spiked_species <- c("Tetragenococcus_halophilus", "Tetragenococcus_sp.")
max_passed_range <- 12 
output_path <- "spike_success_report.docx"

# Convert the phyloseq object to absolute counts
# using scaling factors (example step)
absolute <- convert_to_absolute_counts(physeq_16SASV, scaling_factors)
physeq_absolute <- absolute$physeq_obj 

# Subset the phyloseq object to exclude blank samples (optional step)
physeq_adjusted <- phyloseq::subset_samples(physeq_absolute, sample.or.blank != "blank")

# Generate the spike success report and calculate summary statistics
summary_stats <- conclusion(physeq_adjusted, 
                            merged_spiked_species, 
                            max_passed_range, 
                            output_path)

# Print the summary statistics
print(summary_stats)
}
}
