% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{convert_categorical_to_factors}
\alias{convert_categorical_to_factors}
\title{Convert Categorical Columns to Factors in Sample Data}
\usage{
convert_categorical_to_factors(ps)
}
\arguments{
\item{ps}{A phy#'}
}
\value{
A phyloseq object with updated sample data.
}
\description{
Convert Categorical Columns to Factors in Sample Data
}
\examples{
\dontrun{
ps <- phyloseq::GlobalPatterns
ps_factor <- convert_categorical_to_factors(ps)
}loseq object.
}
