% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{convert_categorical_to_factors}
\alias{convert_categorical_to_factors}
\title{Convert Categorical Columns to Factors in Sample Data}
\usage{
convert_categorical_to_factors(obj)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object containing microbial data.}
}
\value{
A phyloseq object with updated sample data.
}
\description{
Convert Categorical Columns to Factors in Sample Data
}
\examples{
data("physeq_16SOTU", package = "DspikeIn")
ps_factor <- convert_categorical_to_factors(physeq_16SOTU)
}
