% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_absolute_counts.R
\name{convert_to_absolute_counts}
\alias{convert_to_absolute_counts}
\title{Convert Relative ASV Counts to Absolute Counts}
\usage{
convert_to_absolute_counts(physeq, scaling_factors, output_dir = NULL)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbial data.}

\item{scaling_factors}{A numeric vector of scaling factors to convert relative counts to absolute counts.}

\item{output_dir}{A character string specifying the directory to save the output file. Default is NULL, which uses the current working directory.}
}
\value{
NULL. The function saves the absolute counts as a CSV file and prints a message indicating the save location.
}
\description{
This function converts the relative ASV counts in a phyloseq object to absolute counts by multiplying 
the ASV counts by provided scaling factors. The resulting absolute counts are saved as a CSV file.
}
\examples{
# Example usage:
# Assume `spiked_16S` is a phyloseq object and `scaling_factors` is a numeric vector of scaling factors.
# convert_to_absolute_counts(spiked_16S, scaling_factors)
}
