% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{create_list}
\alias{create_list}
\title{Create a List from a Phyloseq or TSE Object}
\usage{
create_list(obj)
}
\arguments{
\item{obj}{A phyloseq or TreeSummarizedExperiment object.}
}
\value{
A list containing the DGE list and updated phyloseq object.
}
\description{
Create a List from a Phyloseq or TSE Object
}
