% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{create_list}
\alias{create_list}
\title{Create a List from a Phyloseq Object}
\usage{
create_list(physeq)
}
\arguments{
\item{physeq}{A phyloseq object.}
}
\value{
A list containing the DGE list and updated phyloseq object.
}
\description{
Create a List from a Phyloseq Object
}
