% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_common_asvs_taxa.R
\name{detect_common_asvs_taxa}
\alias{detect_common_asvs_taxa}
\title{Detect Common ASVs and Taxa from Multiple Phyloseq Objects}
\usage{
detect_common_asvs_taxa(
  phyloseq_list,
  output_common_asvs_csv = "common_asvs.csv",
  output_common_asvs_rds = "common_asvs.rds",
  output_common_taxa_csv = "common_taxa.csv",
  output_common_taxa_rds = "common_taxa.rds",
  return_as_df = FALSE
)
}
\arguments{
\item{phyloseq_list}{A list of phyloseq objects.}

\item{output_common_asvs_csv}{A character string specifying the path to save the common ASVs as a CSV file. Default is "common_asvs.csv".}

\item{output_common_asvs_rds}{A character string specifying the path to save the common ASVs as an RDS file. Default is "common_asvs.rds".}

\item{output_common_taxa_csv}{A character string specifying the path to save the common taxa as a CSV file. Default is "common_taxa.csv".}

\item{output_common_taxa_rds}{A character string specifying the path to save the common taxa as an RDS file. Default is "common_taxa.rds".}

\item{return_as_df}{A logical indicating whether to return the results as data frames. Default is FALSE, meaning the results are returned as phyloseq objects.}
}
\value{
A list containing the phyloseq objects or data frames of common ASVs and common taxa.
}
\description{
This function identifies the common ASVs (Amplicon Sequence Variants) and common taxa detected across multiple phyloseq objects.
It extracts ASVs and taxa from the provided phyloseq objects, finds the common ones, and optionally saves the results to CSV and RDS files.
}
\examples{
# Example usage:
# results <- detect_common_asvs_taxa(list(physeq1, physeq2, physeq3), return_as_df = TRUE)
# common_asvs_df <- results$common_asvs
# common_taxa_df <- results$common_taxa
}
