% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Startup message for DspikeIn package}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{The library name.}

\item{pkgname}{The package name.}
}
\description{
This function displays a startup message when the DspikeIn package is loaded.
}
\keyword{internal}
