% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R, R/adjust_abundance_one_third.R,
%   R/proportion_adj.R, R/summ_ASV_OTUID.R
\name{get_otu_table}
\alias{get_otu_table}
\title{Extract OTU Table from Object}
\usage{
get_otu_table(obj)

get_otu_table(obj)

get_otu_table(obj)

get_otu_table(obj)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object.}
}
\value{
A matrix containing OTU count data.

A matrix containing OTU count data.

A matrix containing OTU count data.

A matrix containing OTU count data.
}
\description{
Retrieves the OTU table from a \code{phyloseq} or \code{TreeSummarizedExperiment} object.

Retrieves the OTU table from a \code{phyloseq} or \code{TreeSummarizedExperiment} object.

Retrieves the OTU table from a \code{phyloseq} or \code{TreeSummarizedExperiment} object.

Retrieves the OTU table from a \code{phyloseq} or \code{TreeSummarizedExperiment} object.
}
