% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_reference_seq}
\alias{get_reference_seq}
\title{Extract Reference Sequences}
\usage{
get_reference_seq(obj)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object.}
}
\value{
A \code{DNAStringSet} object containing the reference sequences.
}
\description{
Retrieves the reference sequences (if available) from an object.
}
