% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_sample_data}
\alias{get_sample_data}
\title{Extract Sample Data}
\usage{
get_sample_data(obj)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object.}
}
\value{
A data frame with sample metadata.
}
\description{
Retrieves the sample metadata.
}
