% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_tax_table}
\alias{get_tax_table}
\title{Extract Taxonomy Table}
\usage{
get_tax_table(obj)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object.}
}
\value{
A data frame containing taxonomy annotations.
}
\description{
Retrieves the taxonomy table from an object.
}
