% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalVariables.R
\name{globalVariables_suppress}
\alias{globalVariables_suppress}
\title{Suppress R CMD check warnings for global variables}
\description{
This file uses `utils::globalVariables()` to declare global variables in order to suppress warnings
from R CMD check about undefined global variables used in non-standard evaluation (NSE).
}
