% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_differential_abundance.R
\name{glom_taxa_at_rank}
\alias{glom_taxa_at_rank}
\title{Glom Taxa at Specific Level}
\usage{
glom_taxa_at_rank(physeq, rank)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{rank}{A character string specifying the taxonomic rank to glom.}
}
\value{
A phyloseq object with taxa glommed at the specified rank.
}
\description{
This function gloms taxa at a specified taxonomic rank.
}
\examples{
ps_glommed <- glom_taxa_at_rank(ps, "Genus")
}
