% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{gm_mean}
\alias{gm_mean}
\title{Helper function for geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical, whether to remove NA values.}
}
\value{
Geometric mean of the input vector.
}
\description{
Helper function for geometric mean
}
