% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{gm_mean}
\alias{gm_mean}
\title{Calculate Geometric Mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical, should missing values (NAs) be removed?}
}
\value{
Geometric mean of x.
}
\description{
Calculate Geometric Mean
}
