% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_differential_abundance.R,
%   R/normalization_set.R
\name{install_and_load}
\alias{install_and_load}
\title{Install and Load Required Packages
This function installs and loads required packages.}
\usage{
install_and_load(packages)

install_and_load(packages)
}
\arguments{
\item{packages}{A character vector of package names to install and load.}
}
\description{
Install and Load Required Packages
This function installs and loads required packages.

-----------------------------------------------------------
}
\examples{
required_packages <- c("phyloseq", "DESeq2", "edgeR", "BiocManager", "BiocGenerics", "ggplot2", "dplyr", "DT")
install_and_load(required_packages)
}
