% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG.R, R/my_custom_theme.R
\docType{data}
\name{MG}
\alias{MG}
\title{MG Color Palette}
\format{
An object of class \code{character} of length 84.

An object of class \code{character} of length 84.
}
\usage{
MG

MG
}
\value{
A character vector of color hex codes.
}
\description{
This function returns a character vector of color palettes used in the package.
}
\keyword{datasets}
