% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG.R, R/microbiome_differential_abundance.R,
%   R/my_custom_theme.R, R/taxa_barplot.R
\name{MG}
\alias{MG}
\title{MG Color Palette}
\usage{
MG()

MG()

MG()

MG()
}
\value{
A character vector of color hex codes.

A character vector of color hex codes.

A character vector of color hex codes.
}
\description{
This function returns a character vector of color palettes used in the package.

This function returns a character vector of color palettes used in the package.
}
\examples{
# Example usage of the MG color palette
my_colors <- MG()
print(my_colors)
}
