% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_shape.R
\docType{data}
\name{MG_shape}
\alias{MG_shape}
\title{Predefined shape vector for consistent plot styling}
\format{
A numeric vector.
}
\usage{
MG_shape
}
\description{
This vector provides predefined shapes for consistent plot styling.
}
\examples{
# Example usage of MG_shape
plot(1:10, pch = MG_shape)
}
\keyword{datasets}
