% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_shapes.R
\docType{data}
\name{MG_shapes}
\alias{MG_shapes}
\title{Predefined Shape Vector for Plot Styling}
\format{
An object of class \code{numeric} of length 26.
}
\usage{
MG_shapes
}
\value{
A numeric vector of shape codes.
}
\description{
A numeric vector of point shapes (0-25) for use in ggplot2.
}
\examples{
# Example usage of MG_shapes in ggplot2
ggplot2::ggplot(mtcars, ggplot2::aes(x = wt, y = mpg, shape = factor(cyl))) +
  ggplot2::geom_point(size = 4) +
  ggplot2::scale_shape_manual(values = MG_shapes[seq_len(3)]) +
  my_custom_theme()

}
\keyword{datasets}
