% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.clr}
\alias{norm.clr}
\title{CLR Normalization (Centered Log-Ratio Transformation)}
\usage{
norm.clr(ps)
}
\arguments{
\item{ps}{A phyloseq object.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
CLR Normalization (Centered Log-Ratio Transformation)
}
