% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.clr}
\alias{norm.clr}
\title{CLR Normalization (Centered Log-Ratio Transformation)}
\usage{
norm.clr(
  ps,
  feature_category = c("all_features", "interquartile_range", "non_zero_counts",
    "specified_features")
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{A character string specifying the feature category ("all_features", "interquartile_range", "non_zero_counts", "specified_features").}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
CLR Normalization (Centered Log-Ratio Transformation)
}
\references{
Aitchison, 1986.
}
