% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.css}
\alias{norm.css}
\title{CSS Normalization (Cumulative Sum Scaling)}
\usage{
norm.css(ps)
}
\arguments{
\item{ps}{A phyloseq object.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
CSS Normalization (Cumulative Sum Scaling)
}
