% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.css}
\alias{norm.css}
\title{CSS Normalization (Cumulative Sum Scaling)}
\usage{
norm.css(ps, feature_category = c("all", "iqlr", "zero", "lvha"))
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{A character string specifying the feature category ("all", "iqlr", "zero", "lvha").}
}
\value{
A phyloseq object with CSS normalized counts.
}
\description{
Assumes equivalent count distributions for low abundant genes across samples up to a threshold. Scales the invariant segment of each sample’s count distribution.
}
\references{
Paulson et al., 2013.
}
