% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.DESeq}
\alias{norm.DESeq}
\title{DESeq Normalization}
\usage{
norm.DESeq(ps, groups, pseudocount = 1)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{groups}{A string specifying the grouping variable in sample data.}

\item{pseudocount}{A numeric value added to avoid zeros in the dataset.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
DESeq Normalization with Pseudocount and Integer Conversion
}
\examples{
\dontrun{
result_DESeq <- normalization_set(ps, method = "DESeq", groups = "SampleType")
normalized_ps_DESeq <- result_DESeq$dat.normed
scaling_factors_DESeq <- result_DESeq$scaling.factor
}
}
