% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.DESeq}
\alias{norm.DESeq}
\title{DESeq Normalization}
\usage{
norm.DESeq(ps, groups)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{groups}{A string specifying the grouping variable in sample data.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
DESeq Normalization
}
\references{
DESeq2 package.
}
