% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.med}
\alias{norm.med}
\title{Median Normalization}
\usage{
norm.med(ps, groups)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{groups}{A vector of group labels.}
}
\value{
A list containing the normalized data and scaling factor.
}
\description{
Median Normalization
}
