% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.Poisson}
\alias{norm.Poisson}
\title{Poisson Normalization and Differential Abundance Function}
\usage{
norm.Poisson(obj, group_var = NULL, pseudocount = 1e-06)
}
\arguments{
\item{obj}{A phyloseq object or matrix of raw counts.}

\item{group_var}{A string specifying the grouping variable in sample data (if phyloseq object).}

\item{pseudocount}{A numeric value added to avoid division by zero.}
}
\value{
A list containing the normalized data, scaling factor, and differential abundance results.
}
\description{
Poisson Normalization and Differential Abundance Function
}
