% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.QN}
\alias{norm.QN}
\title{Quantile Normalization (QN) for phyloseq object}
\usage{
norm.QN(ps, filter = FALSE)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{filter}{Logical, whether to filter low counts.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
Quantile Normalization (QN) for phyloseq object
}
\examples{
\dontrun{
ps <- phyloseq::GlobalPatterns
result_QN <- norm.QN(ps)
normalized_ps_QN <- result_QN$dat.normed
}
}
