% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.rar}
\alias{norm.rar}
\title{Rarefying}
\usage{
norm.rar(ps)
}
\arguments{
\item{ps}{A phyloseq object.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
Rarefying
}
\examples{
\dontrun{
ps <- phyloseq::GlobalPatterns
result_rar <- norm.rar(ps)
normalized_ps_rar <- result_rar$dat.normed
}
}
