% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.rar}
\alias{norm.rar}
\title{Rarefying}
\usage{
norm.rar(obj)
}
\arguments{
\item{obj}{A Phyloseq or TreeSummarizedExperiment objects.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
Rarefying
}
