% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.rar}
\alias{norm.rar}
\title{Rarefying}
\usage{
norm.rar(ps, feature_category = c("all", "iqlr", "zero", "lvha"))
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{A character string specifying the feature category ("all", "iqlr", "zero", "lvha").}
}
\value{
A list containing the normalized data and scaling factor.
}
\description{
Randomly removes reads from different samples until they all have the same predefined number of reads, ensuring equal library size.
}
\references{
McMurdie and Holmes, 2014.
}
