% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.rar}
\alias{norm.rar}
\title{Rarefying}
\usage{
norm.rar(
  ps,
  feature_category = c("all_features", "interquartile_range", "non_zero_counts",
    "specified_features")
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{A character string specifying the feature category ("all_features", "interquartile_range", "non_zero_counts", "specified_features").}
}
\value{
A list containing the normalized data and scaling factor.
}
\description{
Rarefying
}
\references{
McMurdie and Holmes, 2014.
}
