% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.rle}
\alias{norm.rle}
\title{RLE Normalization (Relative Log Expression)}
\usage{
norm.rle(
  ps,
  locfunc = stats::median,
  type = c("poscounts", "ratio"),
  geo_means = NULL,
  control_genes = NULL
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{locfunc}{A function to compute the location statistic (default is median).}

\item{type}{A character string specifying the type of normalization ("poscounts" or "ratio").}

\item{geo_means}{A vector of geometric means for each feature.}

\item{control_genes}{A vector of control genes.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
RLE Normalization (Relative Log Expression)
}
