% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.rle}
\alias{norm.rle}
\title{RLE Normalization (Relative Log Expression)}
\usage{
norm.rle(
  object,
  locfunc = stats::median,
  type = c("poscounts", "ratio"),
  geo_means = NULL,
  control_genes = NULL
)
}
\arguments{
\item{object}{A phyloseq or otu_table object.}

\item{locfunc}{A function to compute the location statistic (default is median).}

\item{type}{A character string specifying the type of normalization ("poscounts" or "ratio").}

\item{geo_means}{A vector of geometric means for each feature.}

\item{control_genes}{A vector of control genes.}
}
\value{
A phyloseq or otu_table object with normalization factors.
}
\description{
Assumes most features are not differential and uses relative abundances to calculate the normalization factor.
}
\references{
Anders and Huber, 2010.
}
