% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.TC}
\alias{norm.TC}
\title{TC Normalization (Total Count Scaling)}
\usage{
norm.TC(obj, groups)
}
\arguments{
\item{obj}{A phyloseq object.}

\item{groups}{A string specifying the grouping variable in sample data.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
TC Normalization (Total Count Scaling)
}
