% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.TMM}
\alias{norm.TMM}
\title{TMM Normalization (Trimmed Mean of M component)}
\usage{
norm.TMM(raw, groups)
}
\arguments{
\item{raw}{A matrix of raw counts.}

\item{groups}{A vector of group labels.}
}
\value{
A list containing the normalized data and scaling factor.
}
\description{
A scaling normalization method used for RNA-seq count data to account for compositional differences between libraries.
}
\references{
Robinson and Oshlack, 2010.
}
