% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.tss}
\alias{norm.tss}
\title{TSS Normalization (Total Sum Scaling)}
\usage{
norm.tss(ps)
}
\arguments{
\item{ps}{A phyloseq object.}
}
\value{
A list containing the normalized phyloseq object and scaling factor
}
\description{
TSS Normalization (Total Sum Scaling)
}
\examples{
\dontrun{
ps <- phyloseq::GlobalPatterns
result_tss <- norm.tss(ps)
normalized_ps_tss <- result_tss$dat.normed
}ctors.
}
