% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.tss}
\alias{norm.tss}
\title{TSS Normalization (Total Sum Scaling)}
\usage{
norm.tss(ps)
}
\arguments{
\item{ps}{A phyloseq object.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
TSS Normalization (Total Sum Scaling)
}
\references{
Bolstad et al., 2003.
}
