% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{norm.UQ}
\alias{norm.UQ}
\title{UQ Normalization (Upper Quartile)}
\usage{
norm.UQ(raw, groups)
}
\arguments{
\item{raw}{A matrix of raw counts.}

\item{groups}{A vector of group labels.}
}
\value{
A list containing the normalized data and scaling factor.
}
\description{
Normalizes data based on the upper quartile of the counts.
}
