% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{normalization_set}
\alias{normalization_set}
\title{the Selected Normalization Method to the Phyloseq Object}
\usage{
normalization_set(obj, method, groups = NULL)
}
\arguments{
\item{obj}{A phyloseq object.}

\item{method}{A character string specifying the normalization method ("TC", "UQ", "med", "DESeq", "Poisson", "QN", "TMM", "clr", "rar", "css", "tss", "rle").}

\item{groups}{A column name of group labels from sample data.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
the Selected Normalization Method to the Phyloseq Object
}
\examples{
\donttest{
if (requireNamespace("DspikeIn", quietly = TRUE)) {
  data("physeq_16SOTU", package = "DspikeIn")
  ps <- physeq_16SOTU

  # TC Normalization
  result_TC <- normalization_set(ps, method = "TC", groups = "Host.species")
  normalized_ps_TC <- result_TC$dat.normed
  scaling_factors_TC <- result_TC$scaling.factor

  # UQ Normalization
  data("physeq_16SOTU", package = "DspikeIn")
  ps <- physeq_16SOTU
  result_UQ <- normalization_set(ps, method = "UQ", groups = "Host.species")
  normalized_ps_UQ <- result_UQ$dat.normed
  scaling_factors_UQ <- result_UQ$scaling.factor

  # Median Normalization
  data("physeq_16SOTU", package = "DspikeIn")
  ps <- physeq_16SOTU
  result_med <- normalization_set(ps, method = "med", groups = "Host.species")
  normalized_ps_med <- result_med$dat.normed
  scaling_factors_med <- result_med$scaling.factor

  # DESeq Normalization
  data("physeq_16SOTU", package = "DspikeIn")
  ps <- physeq_16SOTU
  ps_n <- remove_zero_negative_count_samples(ps)
  result_DESeq <- normalization_set(ps_n, method = "DESeq", groups = "Animal.type")
  normalized_ps_DESeq <- result_DESeq$dat.normed
  scaling_factors_DESeq <- result_DESeq$scaling.factor

  # Poisson Normalization
  data("physeq_16SOTU", package = "DspikeIn")
  ps <- physeq_16SOTU
  result_Poisson <- normalization_set(ps, method = "Poisson", groups = "Host.genus")
  normalized_ps_Poisson <- result_Poisson$dat.normed
  scaling_factors_Poisson <- result_Poisson$scaling.factor

  # Quantile Normalization
  data("physeq_16SOTU", package = "DspikeIn")
  ps <- physeq_16SOTU
  result_QN <- normalization_set(ps, method = "QN")
  normalized_ps_QN <- result_QN$dat.normed
  scaling_factors_QN <- result_QN$scaling.factor

  # TMM Normalization
  data("physeq_16SOTU", package = "DspikeIn")
  ps <- physeq_16SOTU
  result_TMM <- normalization_set(ps, method = "TMM", groups = "Animal.type")
  normalized_ps_TMM <- result_TMM$dat.normed
  scaling_factors_TMM <- result_TMM$scaling.factor

  # CLR Normalization
  data("physeq_16SOTU", package = "DspikeIn")
  ps <- physeq_16SOTU
  result_clr <- normalization_set(ps, method = "clr")
  normalized_ps_clr <- result_clr$dat.normed
  scaling_factors_clr <- result_clr$scaling.factor

  # Rarefying
  data("physeq_16SOTU", package = "DspikeIn")
  ps <- physeq_16SOTU
  result_rar <- normalization_set(ps, method = "rar")
  normalized_ps_rar <- result_rar$dat.normed
  scaling_factors_rar <- result_rar$scaling.factor

  # CSS Normalization
  data("physeq_16SOTU", package = "DspikeIn")
  ps <- physeq_16SOTU
  result_css <- normalization_set(ps, method = "css")
  normalized_ps_css <- result_css$dat.normed
  scaling_factors_css <- result_css$scaling.factor

  # TSS Normalization
  data("physeq_16SOTU", package = "DspikeIn")
  ps <- physeq_16SOTU
  result_tss <- normalization_set(ps, method = "tss")
  normalized_ps_tss <- result_tss$dat.normed
  scaling_factors_tss <- result_tss$scaling.factor

  # RLE Normalization
  data("physeq_16SOTU", package = "DspikeIn")
  ps <- physeq_16SOTU
  result_rle <- normalization_set(ps, method = "rle")
  normalized_ps_rle <- result_rle$dat.normed
  scaling_factors_rle <- result_rle$scaling.factor
}
}

}
