% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{normalization_set}
\alias{normalization_set}
\title{Normalization set function}
\usage{
normalization_set(ps, method, groups = NULL)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{method}{A character string specifying the normalization method ("TC", "UQ", "med", "DESeq", "PoissonSeq", "QN", "SVA", "RUVg", "RUVs", "RUVr", "TMM", "clr", "rar", "css", "tss", "rle").}

\item{groups}{A vector of group labels.}
}
\value{
A list containing the normalized phyloseq object and scaling factors.
}
\description{
Apply the selected normalization method to the phyloseq object.
}
