% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_and_visualize_DA.R
\name{perform_and_visualize_DA}
\alias{perform_and_visualize_DA}
\title{Perform and Visualize Differential Abundance Analysis with edgeR or DESeq2}
\usage{
perform_and_visualize_DA(
  obj,
  method,
  group_var,
  contrast,
  pseudocount = 1,
  significance_level = 0.05,
  output_csv_path = NULL,
  target_glom = "Genus",
  palette = c("#FFEB3B", "#073B4C")
)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object.}

\item{method}{A string: \code{"edgeR"} or \code{"DESeq2"}.}

\item{group_var}{A string: grouping variable in the sample metadata.}

\item{contrast}{A character vector: levels to compare (e.g., \code{c("Control", "Treated")}).}

\item{pseudocount}{A numeric: pseudocount for zero handling (default = \code{1}).}

\item{significance_level}{A numeric: FDR threshold (default = \code{0.05}).}

\item{output_csv_path}{A string: output path for results CSV (optional).}

\item{target_glom}{A string: taxonomic rank to aggregate taxa (default = \code{"Genus"}).}

\item{palette}{A vector: colors for significant and non-significant points (default = \code{c("#FFEB3B", "#073B4C")}).}
}
\value{
A list with:
\item{results}{Data frame of differential abundance results.}
\item{obj_significant}{Filtered \code{phyloseq} or \code{TreeSummarizedExperiment} object.}
\item{plot}{\code{ggplot2} volcano plot with factor levels.}
}
\description{
Performs differential abundance analysis using edgeR or DESeq2, processes data, and creates a volcano plot.
}
\examples{
\donttest{
if (requireNamespace("DspikeIn", quietly = TRUE)) {
  data("physeq_16SOTU", package = "DspikeIn")

  results_edgeR <- perform_and_visualize_DA(
    obj = physeq_16SOTU,
    method = "edgeR",
    group_var = "Diet",
    contrast = c("Insectivore", "Carnivore"),
    output_csv_path = "DA_edgeR.csv",
    target_glom = "Genus",
    significance_level = 0.05
  )

  # Print and visualize results
  print(results_edgeR$plot)
  head(results_edgeR$results)
  print(results_edgeR$obj_significant)

  # Example with a TreeSummarizedExperiment (TSE) object
  tse_16SOTU <- convert_phyloseq_to_tse(physeq_16SOTU)

  results_DESeq2 <- perform_and_visualize_DA(
    obj = tse_16SOTU,
    method = "DESeq2",
    group_var = "Diet",
    contrast = c("Insectivore", "Carnivore"),
    output_csv_path = "DA_DESeq2.csv",
    target_glom = "Genus",
    significance_level = 0.05
  )

  # Print and visualize results
  print(results_DESeq2$plot)
  head(results_DESeq2$results)
  print(results_DESeq2$obj_significant)
}
}
}
