% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_differential_abundance.R
\name{perform_and_visualize_DA}
\alias{perform_and_visualize_DA}
\title{Perform and Visualize Differential Abundance Analysis with edgeR or DESeq2}
\usage{
perform_and_visualize_DA(
  ps,
  method,
  group_var,
  contrast,
  pseudocount = 1,
  significance_level = 0.05,
  output_csv_path = "DA_results.csv",
  point_size = 3,
  target_glom = "Genus",
  palette = color_palette$extended_palette
)
}
\arguments{
\item{ps}{A `phyloseq` object containing microbiome data.}

\item{method}{A character string specifying the method to use: "edgeR" or "DESeq2".}

\item{group_var}{A string specifying the grouping variable in sample data.}

\item{contrast}{A vector specifying the levels to compare (e.g., c("Control", "Treated")).}

\item{pseudocount}{A numeric value for pseudocount addition to handle zeros. Default is 1.}

\item{significance_level}{A numeric value specifying the significance threshold (FDR). Default is 0.05.}

\item{output_csv_path}{Path to save the results as a CSV file. Default is "DA_results.csv".}

\item{point_size}{A numeric value specifying the size of points in the volcano plot. Default is 3.}

\item{target_glom}{A string specifying the taxonomic rank to aggregate taxa. Default is "Genus".}

\item{palette}{A character vector of color hex codes for plotting. Default is `color_palette$extended_palette`.}
}
\value{
A list containing the analysis results, a phyloseq object with significant OTUs/ASVs, and a volcano plot.
}
\description{
This function performs differential abundance analysis using either edgeR or DESeq2 
and generates a volcano plot based on microbiome data.
}
\examples{
\dontrun{
# Using edgeR
results_edgeR <- perform_and_visualize_DA(
  ps = ps,
  method = "edgeR",
  group_var = "Treatment",
  contrast = c("Control", "Treated"),
  output_csv_path = "DA_edgeR.csv",
  target_glom = "Genus",
  significance_level = 0.05
)

# Using DESeq2
results_DESeq2 <- perform_and_visualize_DA(
  ps = ps,
  method = "DESeq2",
  group_var = "Treatment",
  contrast = c("Control", "Treated"),
  output_csv_path = "DA_DESeq2.csv",
  target_glom = "Genus",
  significance_level = 0.05
)

# Visualize Volcano Plot
print(results_edgeR$plot)
ps_sig <-results_edgeR$ps_significant
print(results_DESeq2$plot)
}

}
