% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_differential_abundance.R
\name{perform_and_visualize_differential_abundance}
\alias{perform_and_visualize_differential_abundance}
\title{Perform and Visualize Differential Abundance Analysis}
\usage{
perform_and_visualize_differential_abundance(
  ps,
  group_var,
  method = "edgeR",
  threshold_percentage = 1e-04,
  threshold_mean_abundance = 1e-05,
  threshold_count = 1,
  threshold_relative_abundance = 1e-05,
  significance_level = 0.05,
  point_size = 2
)
}
\arguments{
\item{ps}{A phyloseq object containing the microbiome data.}

\item{group_var}{A string specifying the grouping variable in sample data.}

\item{method}{A string specifying the differential abundance method ("edgeR", "DESeq2", "ANCOMBC"). Default is "edgeR".}

\item{threshold_percentage}{A numeric value for filtering threshold percentage. Default is 0.0001.}

\item{threshold_mean_abundance}{A numeric value for filtering mean abundance. Default is 0.00001.}

\item{threshold_count}{A numeric value for filtering count. Default is 1.}

\item{threshold_relative_abundance}{A numeric value for filtering relative abundance. Default is 0.00001.}

\item{significance_level}{A numeric value specifying the significance level for filtering significant OTUs. Default is 0.05.}

\item{point_size}{A numeric value specifying the size of points in the volcano plot. Default is 2.}
}
\value{
A list containing the final results, the phyloseq object with significant OTUs, and ggplot objects for the visualizations.
}
\description{
This function normalizes and filters the data, performs differential abundance analysis, extracts significant OTUs,
rebuilds the phyloseq object, merges results with metadata, and visualizes the results.
}
\examples{
ps <- remove_zero_negative_count_samples(ps)
results <- perform_and_visualize_differential_abundance(ps, "Animal.ecomode", method = "DESeq2", threshold_percentage = 0.001, threshold_mean_abundance = 0.001, threshold_count = 1, threshold_relative_abundance = 0.001, significance_level = 0.05, point_size = 2)
print(results$plot)
print(results$barplot)
}
