% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_differential_abundance.R
\name{perform_DESeq2}
\alias{perform_DESeq2}
\title{Perform DESeq2 Analysis}
\usage{
perform_DESeq2(ps, group_var, contrast, significance_level = 0.05)
}
\description{
Perform DESeq2 Analysis
}
