% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_differential_abundance.R
\name{perform_edgeR}
\alias{perform_edgeR}
\title{Perform edgeR Analysis}
\usage{
perform_edgeR(ps, group_var, contrast, significance_level = 0.05)
}
\description{
Perform edgeR Analysis
}
