% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq.R
\docType{data}
\name{physeq}
\alias{physeq}
\title{General Phyloseq Object}
\format{
A phyloseq object with taxonomic and abundance data.
}
\source{
Simulated dataset for microbiome analysis.
}
\usage{
data(physeq)
}
\description{
A phyloseq object containing microbiome data for general use.
}
\examples{
data(physeq)
physeq
}
\keyword{datasets}
