% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq_16SOTU.R
\docType{data}
\name{physeq_16SOTU}
\alias{physeq_16SOTU}
\title{Example Phyloseq Object for 16S OTUs}
\format{
A phyloseq object with taxonomic and abundance data.
}
\source{
internal dataset for microbiome analysis.
}
\usage{
data(physeq_16SOTU)
}
\description{
A phyloseq object containing 16S rRNA gene amplicon sequencing data.
}
\examples{
data(physeq_16SOTU)
physeq_16SOTU
}
\keyword{datasets}
