% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq_ITSOTU.R
\docType{data}
\name{physeq_ITSOTU}
\alias{physeq_ITSOTU}
\title{Example Phyloseq Object for ITS OTUs}
\format{
A phyloseq object with taxonomic and abundance data.
}
\source{
internal dataset for microbiome analysis.
}
\usage{
data(physeq_ITSOTU)
}
\description{
A phyloseq object containing ITS amplicon sequencing data.
}
\examples{
data(physeq_ITSOTU)
physeq_ITSOTU
}
\keyword{datasets}
