% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_core_microbiome_custom.R
\name{plot_core_microbiome_custom}
\alias{plot_core_microbiome_custom}
\title{Plot Core Microbiome Prevalence Heatmap}
\usage{
plot_core_microbiome_custom(
  physeq = NULL,
  taxrank = "Genus",
  select_taxa = NULL,
  prevalences = seq(0.01, 1, 0.01),
  detections = 10^seq(log10(0.001), log10(0.3), length = 5),
  min_prevalence = 0.3,
  output_core_csv = NULL,
  output_core_rds = NULL
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbial data.}

\item{taxrank}{A character string specifying the taxonomic rank to glom taxa. Default is "Genus".}

\item{select_taxa}{A character vector of taxa to select. Default is NULL, meaning no specific taxa are selected.}

\item{prevalences}{A numeric vector specifying the prevalence thresholds for plotting. Default is seq(0.01, 1, 0.01).}

\item{detections}{A numeric vector specifying the detection thresholds for plotting. Default is 10^seq(log10(1e-3), log10(0.3), length = 5).}

\item{min_prevalence}{A numeric value specifying the minimum prevalence threshold for core microbiome. Default is 0.3.}

\item{output_core_csv}{A character string specifying the path to save the core microbiome subset as a CSV file. Default is NULL, meaning no CSV file is saved.}

\item{output_core_rds}{A character string specifying the path to save the core microbiome subset as an RDS file. Default is NULL, meaning no RDS file is saved.}
}
\value{
A ggplot2 object representing the core microbiome prevalence heatmap.
}
\description{
This function generates a prevalence heatmap of the core microbiome at a specified taxonomic rank.
It includes a detection threshold for prevalence, performs glomming, pruning, and proportion transformation,
and renames taxa with an ASV prefix. The function requires the `phyloseq`, `microbiomeutilities`, `microbiome`, and `ggplot2` packages.
}
\examples{
\dontrun{
# Example usage:
# Specify a custom detection threshold range
custom_detections <- 10^seq(log10(3e-1), log10(0.5), length = 5)

# Plot the core microbiome using the custom detection thresholds
# and save the results to .rds and .csv files
plot_result <- plot_core_microbiome_custom(physeq_16SASV, 
                                           detections = custom_detections, 
                                           taxrank = "Family", 
                                           output_core_rds = "core_microbiome.rds", 
                                           output_core_csv = "core_microbiome.csv")
}
}
