% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glommed_tree.R
\name{plot_glommed_tree}
\alias{plot_glommed_tree}
\title{Plot a Phylogenetic Tree with Glommed OTUs}
\source{
Uses speedyseq::tree_glom() and ggtree for tree visualization.
}
\usage{
plot_glommed_tree(
  physeq,
  resolution,
  output_prefix = "glommed_tree",
  width = 20,
  height = 18,
  print_plot = TRUE
)
}
\arguments{
\item{physeq}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object containing the phylogenetic data.}

\item{resolution}{A numeric value specifying the resolution for glomming the OTUs.}

\item{output_prefix}{A character string specifying the prefix for the output files. Default is \code{"glommed_tree"}.}

\item{width}{A numeric value specifying the width of the output plot. Default is \code{20}.}

\item{height}{A numeric value specifying the height of the output plot. Default is \code{18}.}

\item{print_plot}{A logical value specifying whether to print the plot. Default is \code{TRUE}.}
}
\value{
\code{NULL}. The function saves the plot to the specified output file.
}
\description{
This function gloms the OTUs in a \code{phyloseq} object at the specified resolution and plots the phylogenetic tree.
If a \code{TreeSummarizedExperiment} (TSE) object is provided, it is first converted to \code{phyloseq}.
}
\examples{
\dontrun{
plot_glommed_tree(Tetragenococcus_tse, resolution = 0.2, output_prefix = "top",
 width = 18, height = 18, print_plot = TRUE)

plot_glommed_tree(Tetragenococcus, resolution = 0.3, output_prefix = "tse_tree")
}
}
