% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glommed_tree.R
\name{plot_glommed_tree}
\alias{plot_glommed_tree}
\title{Plot a Phylogenetic Tree with Glommed OTUs}
\usage{
plot_glommed_tree(
  physeq,
  resolution,
  output_prefix = "glommed_tree",
  width = 20,
  height = 18,
  print_plot = TRUE
)
}
\arguments{
\item{physeq}{A phyloseq object containing the phylogenetic data.}

\item{resolution}{A numeric value specifying the resolution for glomming the OTUs.}

\item{output_prefix}{A character string specifying the prefix for the output files. Default is "glommed_tree".}

\item{width}{A numeric value specifying the width of the output plot. Default is 18.}

\item{height}{A numeric value specifying the height of the output plot. Default is 18.}

\item{print_plot}{A logical value specifying whether to print the plot. Default is TRUE.}
}
\value{
NULL. The function saves the plot to the specified output file.
}
\description{
This function gloms the OTUs in the phyloseq object at the specified resolution and plots the tree.
}
\examples{
\dontrun{
if (interactive()) {
  # Plot the tree with glommed OTUs
  plot_glommed_tree(Tetragenococcus, resolution = 0.2, 
  output_prefix = "top", width = 18, height = 18)
}
}
}
