% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_custom.R
\name{plot_tree_custom}
\alias{plot_tree_custom}
\title{Plot and Save a Phylogenetic Tree (phyloseq/TSE)}
\usage{
plot_tree_custom(
  obj,
  output_prefix = "phy_tree",
  width = 10,
  height = 10,
  layout = "circular",
  tip_labels = TRUE,
  metadata_column = NULL,
  print_plot = TRUE
)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object containing phylogenetic data.}

\item{output_prefix}{A character string specifying the prefix for the output files. Default is \code{"phy_tree"}.}

\item{width}{A numeric value specifying the width of the output plot (in inches). Default is \code{10}.}

\item{height}{A numeric value specifying the height of the output plot (in inches). Default is \code{10}.}

\item{layout}{A character string specifying the tree layout. Options are \code{"rectangular"}, \code{"circular"}, \code{"fan"}, and \code{"radial"}. Default is \code{"circular"}.}

\item{tip_labels}{A logical indicating whether to display tip labels. Default is \code{TRUE}.}

\item{metadata_column}{A string specifying a column from \code{colData(obj)} (TSE) or \code{sample_data(obj)} (phyloseq) to annotate the tree nodes. Default is \code{NULL}.}

\item{print_plot}{A logical indicating whether to print the plot to the console. Default is \code{TRUE}.}
}
\value{
NULL. Saves the tree plot as PNG and PDF files.
}
\description{
Plots a phylogenetic tree from either a \code{phyloseq} or \code{TreeSummarizedExperiment} object.
If the tree is missing in a TSE, it converts the object to \code{phyloseq} before plotting.
}
\examples{
\dontrun{
if (requireNamespace("DspikeIn", quietly = TRUE)) {
  data("physeq_16SOTU", package = "DspikeIn")

  # Plot tree for phyloseq object
  plot_tree_custom(physeq_16SOTU, output_prefix = "phylo_plot", layout = "circular")

  tse_16SOTU <- convert_phyloseq_to_tse(physeq_16SOTU)

  # Example with a TreeSummarizedExperiment object and metadata annotation
  plot_tree_custom(
    tse_16SOTU,
    output_prefix = "tse_plot",
    layout = "fan",
    metadata_column = "SampleType"
  )
}
}
}
