% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_custom.R
\name{plot_tree_custom}
\alias{plot_tree_custom}
\title{Plot and Save a Phylogenetic Tree}
\usage{
plot_tree_custom(
  physeq,
  output_prefix = "phy_tree",
  width = 10,
  height = 10,
  layout = "rectangular",
  print_plot = TRUE
)
}
\arguments{
\item{physeq}{A phyloseq object containing the phylogenetic data.}

\item{output_prefix}{A character string specifying the prefix for the output files. Default is "phy_tree".}

\item{width}{A numeric value specifying the width of the output plot. Default is 10.}

\item{height}{A numeric value specifying the height of the output plot. Default is 10.}

\item{layout}{A character string specifying the layout of the tree. Options are "rectangular", "circular", "fan", and "radial". Default is "rectangular".}

\item{print_plot}{A logical value specifying whether to print the plot. Default is TRUE.}
}
\value{
NULL. The function saves the plot to the specified output file.
}
\description{
This function plots a phylogenetic tree from a phyloseq object and saves the plot to a specified file.
}
\examples{
# Plot and save the phylogenetic tree
plot_tree_custom(Tetragenococcus, output_prefix = "p0", width = 18, height = 18, layout = "circular")
}
