% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_with_alignment.R
\name{plot_tree_with_alignment}
\alias{plot_tree_with_alignment}
\title{Plot a Phylogenetic Tree with Multiple Sequence Alignment}
\usage{
plot_tree_with_alignment(
  physeq,
  output_prefix = "physeq_tree_alignment",
  width = 15,
  height = 15
)
}
\arguments{
\item{physeq}{A phyloseq object containing the phylogenetic data.}

\item{output_prefix}{A character string specifying the prefix for the output files. Default is "physeq_tree_alignment".}

\item{width}{A numeric value specifying the width of the output plot. Default is 15.}

\item{height}{A numeric value specifying the height of the output plot. Default is 15.}
}
\value{
NULL. The function saves the plots to the specified output files.
}
\description{
This function generates a phylogenetic tree plot with an additional plot of the multiple sequence alignment.
}
\examples{
if (interactive()) {
  # Plot the phylogenetic tree with multiple sequence alignment
  plot_tree_with_alignment(Tetragenococcus, 
  output_prefix = "tree_alignment", width = 15, height = 15)
}
}
