% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbar_abundance.R
\name{plotbar_abundance}
\alias{plotbar_abundance}
\title{Plot Bar Abundance for Phyloseq Data}
\usage{
plotbar_abundance(
  physeq,
  level = "Genus",
  color = NULL,
  group = NULL,
  top = 20,
  return = FALSE,
  x_size = 12,
  y_size = 12,
  legend_key_size = 1.5,
  legend_text_size = 12,
  legend_nrow = 20,
  relativize = TRUE,
  output_prefix = NULL
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbiome data.}

\item{level}{A character string specifying the taxonomic level to plot (e.g., "Genus", "Family").}

\item{color}{A character vector specifying colors to use for the different taxa. Default is NULL.}

\item{group}{A character string specifying the grouping variable. Default is NULL.}

\item{top}{An integer specifying the number of top taxa to display. Default is 20.}

\item{return}{A logical indicating whether to return the summarized data frame. Default is FALSE.}

\item{x_size}{An integer specifying the size of the x-axis text. Default is 12.}

\item{y_size}{An integer specifying the size of the y-axis text. Default is 12.}

\item{legend_key_size}{A numeric value specifying the size of the legend keys. Default is 1.5.}

\item{legend_text_size}{An integer specifying the size of the legend text. Default is 12.}

\item{legend_nrow}{An integer specifying the number of rows for the legend. Default is 20.}

\item{relativize}{A logical indicating whether to relativize the data. Default is TRUE.}

\item{output_prefix}{A character string specifying the prefix for the output file names. Default is NULL.}
}
\value{
A ggplot object if return is FALSE, otherwise a data frame with summarized data.
}
\description{
This function generates bar plots for phyloseq data at a specified taxonomic level,
with options to customize the appearance, relativize the data, and save the plots.
}
\examples{
# Plot relativized abundance or select relativize = FALSE to plot absolute abundance
plot <- plotbar_abundance(physeq_16S_adj_scaled_absolute_abundance, level = "Family", 
group = "Env_broad_scale.x", top = 10, x_size = 10, y_size = 10, legend_key_size = 2, legend_text_size = 14,
legend_nrow = 10, relativize = TRUE, output_prefix = "relativized_abundance_plot")
print(plot)

# Plot non-relativized (absolute) abundance
plot_absolute <- plotbar_abundance(physeq_16S_adj_scaled_absolute_abundance, 
level = "Family", group = "Env_broad_scale.x", top = 10, x_size = 10, y_size = 10, 
legend_key_size = 2, legend_text_size = 14, legend_nrow = 10, relativize = FALSE, 
output_prefix = "non_relativized_abundance_plot")
print(plot_absolute)

# Return summarized data
data <- plotbar_abundance(physeq_16S_adj_scaled_absolute_abundance, level = "Family", 
group = "Env_broad_scale.x", top = 10, return = TRUE)
}
