% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre_processing_hashcodes.R
\name{Pre_processing_hashcodes}
\alias{Pre_processing_hashcodes}
\title{Pre-process phyloseq or TSE object based on hashcodes}
\usage{
Pre_processing_hashcodes(
  obj,
  hashcodes,
  merge_method = c("sum", "max"),
  output_prefix = "merged_physeq"
)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object.}

\item{hashcodes}{A character vector of taxon hashcodes (OTU row names).}

\item{merge_method}{The method to merge taxa: "sum" or "max".}

\item{output_prefix}{A prefix for the output file names.}
}
\value{
A processed phyloseq or TSE object.
}
\description{
Subsets, merges, and saves taxa based on hashcodes and a specified merge method ("sum" or "max").
This function pre-processes a \code{phyloseq} or \code{TreeSummarizedExperiment} (TSE) object by subsetting, merging,
and saving taxa based on provided hashcodes. It retains taxonomic information and creates intermediate datasets
for further downstream analysis.
}
\examples{
if (requireNamespace("DspikeIn", quietly = TRUE)) {
  data("physeq_16SOTU", package = "DspikeIn")

  # Subset phyloseq object to include only Tetragenococcus species
  Tetragenococcus <- phyloseq::subset_taxa(
    physeq_16SOTU,
    Species == "Tetragenococcus_halophilus" | Species == "Tetragenococcus_sp."
  )

  # Extract hashcodes (OTU IDs)
  hashcodes <- row.names(phyloseq::otu_table(Tetragenococcus))

  # Temp file paths
  sum_prefix <- file.path(tempdir(), "merged_physeq_sum")
  max_prefix <- file.path(tempdir(), "merged_physeq_max")

  # Process data using the "sum" merging method
  processed_data_sum <- Pre_processing_hashcodes(
    physeq_16SOTU,
    hashcodes,
    merge_method = "sum")

  # Convert to TreeSummarizedExperiment (TSE)
  Tetragenococcus_TSE <- convert_phyloseq_to_tse(Tetragenococcus)

  # Extract hashcodes for TSE
  hashcodes <- rownames(get_otu_table(Tetragenococcus_TSE))

  # Convert full phyloseq object to TSE format
  tse_16SOTU <- convert_phyloseq_to_tse(physeq_16SOTU)

  # Load precomputed TSE object (optional)
  data("tse_16SOTU", package = "DspikeIn")

  # Process data using the "max" merging method
  processed_data_max <- Pre_processing_hashcodes(
    tse_16SOTU,
    hashcodes,
    merge_method = "max" )

  # Clean up
  files_to_delete <- paste0(c(sum_prefix, max_prefix), "_processed.rds")
  invisible(file.remove(files_to_delete[file.exists(files_to_delete)]))
}
}
