% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre_processing_hashcodes.R
\name{Pre_processing_hashcodes}
\alias{Pre_processing_hashcodes}
\title{Pre-process phyloseq object based on hashcodes}
\usage{
Pre_processing_hashcodes(
  spiked_16S,
  hashcodes,
  merge_method = c("sum", "max"),
  output_prefix = "merged_physeq"
)
}
\arguments{
\item{spiked_16S}{A phyloseq object containing the spiked dataset.}

\item{hashcodes}{A character vector of hashcodes (row names of the OTU table) to be processed.}

\item{merge_method}{The method to use for merging taxa: "sum" or "max".}

\item{output_prefix}{A character string to be used as a prefix for the output file names.}
}
\value{
A phyloseq object with the processed data.
}
\description{
This function pre-processes a phyloseq object by subsetting, merging, and saving taxa
based on specified hashcodes and a merge method (sum or max). It retains the taxa 
information and creates various intermediate and final datasets for further processing.
}
\examples{
Tetra <- subset_taxa(physeqASV16, Species == "Tetragenococcus_halophilus" | Species == "Tetragenococcus_sp")
hashcodes <- row.names(phyloseq::otu_table(Tetra))
processed_data_sum <- Pre_processing_hashcodes(physeqASV16, hashcodes, 
merge_method = "sum", output_prefix = "merged_physeq_sum")
processed_data_max <- Pre_processing_hashcodes(physeqASV16, hashcodes, 
merge_method = "max", output_prefix = "merged_physeq_max")
summ_count_phyloseq(processed_data_sum)
}
