% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre_processing_species.R
\name{Pre_processing_species}
\alias{Pre_processing_species}
\title{Pre-process species in a phyloseq or TSE object by merging ASVs/OTUs}
\usage{
Pre_processing_species(
  obj,
  species_name,
  merge_method = c("sum", "max"),
  output_file = NULL
)
}
\arguments{
\item{obj}{A \code{phyloseq} or \code{TreeSummarizedExperiment} object.}

\item{species_name}{A character vector of species names to merge.}

\item{merge_method}{\code{"sum"} (default) or \code{"max"}: method for merging ASV counts.}

\item{output_file}{Optional file path to save the processed object.}
}
\value{
A processed \code{phyloseq} or \code{TreeSummarizedExperiment} object.
}
\description{
Merges ASVs/OTUs while ensuring the phylogenetic tree and reference sequences remain intact.
If tree or refseq become mismatched, they are pruned or removed safely.
}
\examples{
\dontrun{
library(DspikeIn)
data("physeq_16SOTU", package = "DspikeIn")
 spiked_cells <- 1847
 species_name <- spiked_species <- c("Tetragenococcus_halophilus", "Tetragenococcus_sp.")

 merged_sum <- Pre_processing_species(physeq_16SOTU, species_name, merge_method = "sum")

# Convert phyloseq to TSE format
tse_16SOTU <- convert_phyloseq_to_tse(physeq_16SOTU)

species_name <- c("Tetragenococcus_halophilus", "Tetragenococcus_sp." )

# Merge species in TSE format
merged_TSE_sum <- Pre_processing_species(
    tse_16SOTU,
    species_name,
    merge_method = "sum",
    output_file = "merged_TSE_sum.rds"
)
}
}
