% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre_processing_species.R
\name{Pre_processing_species}
\alias{Pre_processing_species}
\title{Pre-process species in a phyloseq object}
\usage{
Pre_processing_species(
  physeq,
  species_name,
  merge_method = c("sum", "max"),
  output_file = NULL
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbiome data.}

\item{species_name}{A character vector of species names to be processed.}

\item{merge_method}{The method to use for merging ASVs: "sum" or "max".}

\item{output_file}{Optional. A file path to save the merged phyloseq object.}
}
\value{
A phyloseq object with the specified species pre-processed.
}
\description{
This function pre-processes species in a phyloseq object by merging ASVs
based on a specified method (sum or max). It retains the Genus and Species
information for the merged taxa.
}
\examples{
if (interactive()) {
  Tetragenococcus <- phyloseq::subset_taxa(physeq_16SASV, 
    Species == "Tetragenococcus_halophilus" | Species == "Tetragenococcus_sp")
  hashcodes <- row.names(phyloseq::otu_table(Tetragenococcus))
  processed_data_sum <- Pre_processing_hashcodes(physeq_16SASV, hashcodes, 
    merge_method = "sum", output_prefix = "merged_physeq_sum")
  processed_data_max <- Pre_processing_hashcodes(physeq_16SASV, hashcodes, 
    merge_method = "max", output_prefix = "merged_physeq_max")
  summ_count_phyloseq(processed_data_sum)
}
}
