% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre_processing_species.R
\name{Pre_processing_species}
\alias{Pre_processing_species}
\title{Pre-process species in a phyloseq object}
\usage{
Pre_processing_species(
  physeq,
  species_name,
  merge_method = c("sum", "max"),
  output_file = NULL
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbiome data.}

\item{species_name}{A character vector of species names to be processed.}

\item{merge_method}{The method to use for merging ASVs: "sum" or "max".}

\item{output_file}{Optional. A file path to save the merged phyloseq object.}
}
\value{
A phyloseq object with the specified species pre-processed.
}
\description{
This function pre-processes species in a phyloseq object by merging ASVs
based on a specified method (sum or max). It retains the Genus and Species
information for the merged taxa.
}
\examples{
Tetragenococcus <- phyloseq::subset_taxa(physeq_ASV16, Genus == "Tetragenococcus")
species_name <- c("Tetragenococcus_halophilus", "Tetragenococcus_sp")
merged_physeq_sum <- Pre_processing_species(physeq_ASV16, species_name, 
merge_method = "sum", output_file = "merged_physeq_sum.rds")
merged_physeq_max <- Pre_processing_species(Tetrasum, species_name, 
merge_method = "max", output_file = "merged_physeq_max.rds")

# Verify the results
summ_count_phyloseq(merged_physeq_sum)
Tetra_check <- phyloseq::subset_taxa(merged_physeq_sum, Genus == "Tetragenococcus")
}
