% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{process_data_with_feature_category}
\alias{process_data_with_feature_category}
\title{Process Data with Feature Category}
\usage{
process_data_with_feature_category(
  ps,
  feature_category = c("all_features", "interquartile_range", "non_zero_counts",
    "specified_features")
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{feature_category}{A character string specifying the feature category ("all_features", "interquartile_range", "non_zero_counts", "specified_features").}
}
\value{
A processed phyloseq object.
}
\description{
Process Data with Feature Category
}
