% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_adj.R
\name{proportion_adj}
\alias{proportion_adj}
\title{Proportionally Adjust Abundance}
\usage{
proportion_adj(physeq, output_file = "proportion_adjusted_physeq.rds")
}
\arguments{
\item{physeq}{A phyloseq object containing the microbiome data.}

\item{output_file}{A character string specifying the output file name for the adjusted phyloseq object. Default is "proportion_adjusted_physeq.rds".}
}
\value{
A phyloseq object with the proportionally adjusted and rounded abundance data.
}
\description{
This function normalizes the abundance data in a phyloseq object by adjusting each sample's counts
based on a total value, typically the maximum total sequence count across all samples. The adjusted
counts are then rounded to the nearest integer.
}
\examples{
# Proportionally adjust the abundance data
normalized_physeq <- proportion_adj(physeq, output_file = "proportion_adjusted_physeq.rds")
}
