% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_subsample_WithReductionFactor.R
\name{random_subsample_WithReductionFactor}
\alias{random_subsample_WithReductionFactor}
\title{Random Subsampling with Reduction Factor}
\usage{
random_subsample_WithReductionFactor(
  physeq,
  reduction_factor = 3,
  output_file = "Less_subsampled_physeq.rds"
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbiome data.}

\item{reduction_factor}{A numeric value specifying the factor by which to reduce the counts. Default is 3.}

\item{output_file}{A character string specifying the output file name for the subsampled phyloseq object. Default is "Less_subsampled_physeq.rds".}
}
\value{
A phyloseq object with the subsampled OTU table.
}
\description{
This function performs random subsampling on the OTU table of a phyloseq object,
reducing the counts of each ASV (Amplicon Sequence Variant) by a specified reduction factor.
The resulting subsampled phyloseq object is saved to a specified output file.
}
\examples{
# Perform random subsampling with a reduction factor of 10
red <- random_subsample_WithReductionFactor(spiked_16S, reduction_factor = 10)
# Summarize the subsampled phyloseq object
summ_phyloseq_sampleID(red)
}
