% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_selected_ASVs.R
\name{RandomForest_selected_ASVs}
\alias{RandomForest_selected_ASVs}
\title{Select ASVs Using Random Forest}
\usage{
RandomForest_selected_ASVs(
  physeq,
  response_var,
  minlib = 15000,
  prunescale = 1e-04,
  ntree = 100,
  n_top_predictors = 50,
  output_csv = "randomforest_results.csv",
  na_vars = NULL
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbiome data.}

\item{response_var}{A character string specifying the response variable from the sample data.}

\item{minlib}{A numeric value specifying the minimum library size. Default is 15000.}

\item{prunescale}{A numeric value specifying the pruning scale for rare OTUs. Default is 0.0001.}

\item{ntree}{An integer specifying the number of trees to grow in the Random Forest. Default is 100.}

\item{n_top_predictors}{An integer specifying the number of top predictors to select. Default is 50.}

\item{output_csv}{A character string specifying the output CSV file name. Default is "randomforest_results.csv".}

\item{na_vars}{A character vector specifying the sample data variables to check for NA values. Default is NULL.}
}
\value{
A pruned phyloseq object containing the selected ASVs.
}
\description{
This function selects the most important ASVs (Amplicon Sequence Variants) based on a Random Forest model.
The selected ASVs are saved to a CSV file.
}
\examples{
# Ensure you have a phyloseq object `physeq`
# Use the name of the column in sample_data for response_var
rf_physeq <- RandomForest_selected_ASVs(physeq_16SOTU,response_var = "Host.genus",
na_vars = c("Habitat", "Ecoregion.III", "Host.genus", "Diet"))
}
