% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomsubsample_Trimmed_evenDepth.R
\name{randomsubsample_Trimmed_evenDepth}
\alias{randomsubsample_Trimmed_evenDepth}
\title{Subsampling to an Equal Sequencing Depth}
\usage{
randomsubsample_Trimmed_evenDepth(
  physeq,
  smalltrim = 0.001,
  replace = TRUE,
  output_file = "randomsubsample_Trimmed_evenDepth.rds"
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbial data.}

\item{smalltrim}{A numeric value specifying the trimming percentage to exclude very low abundant taxa. Default is 0.001.}

\item{replace}{A logical value indicating whether to sample with replacement. Default is TRUE.}

\item{output_file}{A character string specifying the output file name for the subsampled phyloseq object. Default is "randomsubsample_Trimmed_evenDepth.rds".}
}
\value{
A phyloseq object with subsampled sequencing depths.
}
\description{
This function performs subsampling to an equal sequencing depth, determined by the sample with the lowest sequencing depth after excluding very low abundant taxa.
It rounds down the result to the nearest integer (floor). Caution: you might lose samples.
}
\examples{
# Perform subsampling
spiked_ITS_evenDepth <- randomsubsample_Trimmed_evenDepth(spiked_ITS, smalltrim = 0.001)
}
