% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{regression_plot}
\alias{regression_plot}
\title{Create a Regression Plot with Faceting by Range}
\usage{
regression_plot(
  data,
  x_var,
  y_var,
  custom_range = c(0.1, 15, 30, 50, 75, 100),
  formula = y ~ x,
  plot_title = NULL,
  pseudocount = 1e-06
)
}
\arguments{
\item{data}{A data frame containing the variables to plot.}

\item{x_var}{A string specifying the name of the x-axis variable.}

\item{y_var}{A string specifying the name of the y-axis variable.}

\item{custom_range}{A numeric vector for defining custom ranges for the 'Percentage' column (default: c(0.1, 15, 30, 50, 75, 100)).}

\item{formula}{A formula for the regression equation (default: y ~ x).}

\item{plot_title}{A string specifying the title of the plot (default: NULL, no title will be shown if not provided).}

\item{pseudocount}{A small numeric value that is added to both the x and y variables
to avoid issues when there are zero values in the data. This is helpful in situations 
where zero values would cause issues with calculations, such as log transformations 
or statistical modeling. Default is 1e-6.}
}
\value{
A ggplot2 object.
}
\description{
This function generates a customizable scatter plot with a linear regression line, 
statistical equation, and facets based on a specified range variable.
}
\examples{
\dontrun{
# Example of how to call the function with custom ranges
plot_object <- regression_plot(
  data = metadata, 
  x_var = "Richness", 
  y_var = "Total_Reads_spiked", 
  custom_range = c(0.1, 15, 30, 50, 75, 100)
)
print(plot_object)
}
}
