% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_differential_abundance.R,
%   R/normalization_set.R
\name{remove_zero_negative_count_samples}
\alias{remove_zero_negative_count_samples}
\title{Remove Samples with Zero, Negative Counts, or NA Values and Add Pseudocount}
\usage{
remove_zero_negative_count_samples(ps, pseudocount = 1e-06)

remove_zero_negative_count_samples(ps, pseudocount = 1e-06)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{pseudocount}{A numeric value to add to avoid zero counts.}
}
\value{
A phyloseq object with updated OTU table after pseudocount addition.

A phyloseq object with filtered and adjusted OTU table.
}
\description{
This function removes samples with zero, negative counts, or NA values and adds pseudocounts.
}
\examples{
\dontrun{
ps <- phyloseq::GlobalPatterns
ps_clean <- remove_zero_negative_count_samples(ps)
}
}
