% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{remove_zero_negative_count_samples}
\alias{remove_zero_negative_count_samples}
\title{Remove samples with zero or negative counts}
\usage{
remove_zero_negative_count_samples(ps)
}
\arguments{
\item{ps}{A phyloseq object.}
}
\value{
A pruned phyloseq object without samples with zero or negative counts.
}
\description{
Remove samples with zero or negative counts
}
