% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{set_nf}
\alias{set_nf}
\title{Set normalization factors}
\usage{
set_nf(object, nf)
}
\arguments{
\item{object}{A phyloseq or otu_table object.}

\item{nf}{A vector of normalization factors.}
}
\value{
The object with normalization factors set.
}
\description{
Set normalization factors
}
