% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R
\name{set_nf}
\alias{set_nf}
\title{Set normalization factors in the sample data of the phyloseq object.}
\usage{
set_nf(ps, scaling.factor)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{scaling.factor}{A vector of normalization factors.}
}
\value{
A phyloseq object with updated sample data.
}
\description{
Set normalization factors in the sample data of the phyloseq object.
}
