% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapiro_Wilk_test.R
\name{Shapiro_Wilk_test}
\alias{Shapiro_Wilk_test}
\title{Perform Shapiro-Wilk Test for Normality with Transformations}
\usage{
Shapiro_Wilk_test(data, grouping_var)
}
\arguments{
\item{data}{A data frame containing the data to test.}

\item{grouping_var}{A character string specifying the column name to use for grouping.}
}
\value{
NULL. The function prints the Shapiro-Wilk test results for each variable and transformation.
}
\description{
This function performs the Shapiro-Wilk test for normality of residuals
on numeric variables within a data frame, using different transformations.
}
\examples{
\donttest{
if (requireNamespace("DspikeIn", quietly = TRUE)) {
  data("methods", package = "DspikeIn")
  # Perform the Shapiro-Wilk test on the dataset
  Shapiro_Wilk_test(methods, "Methods")
}
}
}
