% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_barplot.R
\name{taxa_barplot}
\alias{taxa_barplot}
\title{Generate a Taxa Barplot}
\usage{
taxa_barplot(
  physeq,
  target_glom = "Genus",
  custom_tax_names = NULL,
  normalize = TRUE,
  treatment_variable = "Treatment",
  abundance_type = "relative",
  x_angle = 20,
  fill_variable = target_glom,
  facet_variable = "Phylum"
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbiome data.}

\item{target_glom}{A character string specifying the taxonomic rank to plot (e.g., "Genus").
This specifies the taxonomic level at which the data should be aggregated and visualized.}

\item{custom_tax_names}{A character vector specifying custom taxonomic names for the levels. Default is NULL.
This allows the user to provide custom names for the taxonomic levels in the taxonomy table of the phyloseq object.}

\item{normalize}{A logical indicating whether to normalize the sample counts to relative abundances. Default is TRUE.}

\item{treatment_variable}{A character string specifying the treatment variable to use for the x-axis. Default is "Treatment".}

\item{abundance_type}{A character string specifying whether to plot "relative" or "absolute" abundance. Default is "relative".}

\item{x_angle}{A numeric value specifying the angle of the x-axis text labels. Default is 20.}

\item{fill_variable}{A character string specifying the variable to use for fill in stacking taxa. Default is target_glom.}

\item{facet_variable}{A character string specifying the variable to use for faceting. Default is "Phylum".}
}
\value{
A list containing the ggplot2 bar plot object (`barplot`) and the pruned phyloseq object (`taxa_data`) with the top taxa.
}
\description{
This function creates a bar plot of the relative or absolute abundances of taxa at a specified taxonomic rank.
The top 20 taxa are selected, and the plot can be customized with various options.
The function uses the `tax_glom` function from the `phyloseq` package for taxonomic grouping.
}
\examples{
# Generate a taxa barplot for the Genus rank with relative abundance
bp <- taxa_barplot(physeqASV16, target_glom = "Genus", normalize = TRUE, treatment_variable = "animal.type", abundance_type = "relative")
print(bp$barplot)

# Generate a taxa barplot for the Genus rank with absolute abundance
bp <- taxa_barplot(physeqASV16, target_glom = "Genus", normalize = FALSE, treatment_variable = "animal.type", abundance_type = "absolute")
print(bp$barplot)
}
