% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_set.R, R/tidy_phyloseq.R
\name{tidy_phyloseq}
\alias{tidy_phyloseq}
\title{Tidy a Phyloseq Object and Remove Zero/Negative Count Samples}
\usage{
tidy_phyloseq(my_phyloseq)

tidy_phyloseq(my_phyloseq)
}
\arguments{
\item{my_phyloseq}{A phyloseq object containing the taxonomic and abundance data.}

\item{pseudocount}{A numeric value to add as a pseudocount.}
}
\value{
A cleaned and tidied phyloseq object.

A cleaned and tidied phyloseq object.
}
\description{
This function cleans and tidies a phyloseq object by:
- Fixing taxa names
- Setting taxonomic ranks
- Trimming whitespace
- Merging unidentified taxa at a specified rank
- Removing taxa with zero counts
- Removing "Chloroplast" and "Mitochondria" classified taxa
- Removing samples with zero, negative counts, or NA values
- Adding a pseudocount to avoid zero counts

This function cleans and tidies a phyloseq object by performing the following steps:
- Fixes taxa names by removing any characters followed by '__' and any spaces after '__'
- Sets taxonomic ranks to standard names and requires 7 ranks
- Trims leading and trailing whitespace from taxa names
- Removes taxa with zero counts across all samples
- Removes taxa classified as "Chloroplast" at the Class level
- Removes taxa classified as "Mitochondria" at the Family level
}
\examples{
\dontrun{
# Example usage:
spiked_16S <- tidy_phyloseq(spiked_16S, pseudocount = 1e-6)
}
\dontrun{
# Example usage:
spiked_16S <- tidy_phyloseq(spiked_16S)
}
}
