% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_plot.R
\name{transform_plot}
\alias{transform_plot}
\title{Generate and Save Boxplots with Statistical Tests}
\usage{
transform_plot(
  data = NULL,
  x_var = NULL,
  y_vars = NULL,
  methods_var = NULL,
  MG = MG,
  adjustment = "holm",
  output_prefix = "plot",
  width = 15,
  height = 13,
  stat_test = "kruskal.test",
  X = NULL,
  Y = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  bcol = "bisque",
  p.adj = "none",
  cexy = 1.5,
  varwidth = TRUE,
  las = 1,
  paired = FALSE
)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{x_var}{A character string specifying the column name for the x variable.}

\item{y_vars}{A character vector specifying the column names for the y variables.}

\item{methods_var}{A character string specifying the column name for the grouping variable.}

\item{MG}{A character vector specifying the colors for the boxplots. Default is MG.}

\item{adjustment}{A character string specifying the method for p-value adjustment. Default is "holm".}

\item{output_prefix}{A character string specifying the prefix for the output file names. Default is "plot".}

\item{width}{A numeric value specifying the width of the output plot. Default is 15.}

\item{height}{A numeric value specifying the height of the output plot. Default is 13.}

\item{stat_test}{A character string specifying the statistical test to use ("kruskal.test" or "anova"). Default is "kruskal.test".}

\item{X}{A numeric vector of data values.}

\item{Y}{A factor or character vector of grouping variables.}

\item{main}{A character string specifying the main title of the plot. Default is NULL.}

\item{xlab}{A character string specifying the x-axis label. Default is NULL.}

\item{ylab}{A character string specifying the y-axis label. Default is NULL.}

\item{bcol}{A character string specifying the box color. Default is "bisque".}

\item{p.adj}{A character string specifying the method for p-value adjustment. Default is "none".}

\item{cexy}{A numeric value specifying the text size for axis labels and titles. Default is 1.5.}

\item{varwidth}{A logical value specifying whether the boxes should have variable widths. Default is TRUE.}

\item{las}{An integer specifying the orientation of axis labels. Default is 1.}

\item{paired}{A logical value specifying whether the data are paired. Default is FALSE.}
}
\value{
A list containing the ggplot2 boxplot objects and the comparison results, or the comparison results and p-value for individual boxplots.
}
\description{
This function generates boxplots for multiple y variables and performs statistical tests (Kruskal-Wallis or ANOVA) to compare groups. It also generates individual boxplots for Y variables and performs statistical tests (Kruskal-Wallis or Wilcoxon) to compare groups, and saves the plots as PNG and PDF files.
}
\examples{
# Example usage for transform_plot:
# y_vars <- c("Spike.percentage", "Total.reads", "Spike.reads")
# transform_plot(data = scaled, x_var = "Methods", y_vars = y_vars, methods_var = "Methods", MG = MG, stat_test = "kruskal.test")
# transform_plot(data = scaled, x_var = "Methods", y_vars = y_vars, methods_var = "Methods", MG = MG, stat_test = "anova")
#
# Example usage for individual boxplots:
# X <- rnorm(100)
# Y <- rep(c("A", "B"), each = 50)
# transform_plot(X = X, Y = Y)
}
