% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tse.R
\docType{data}
\name{tse}
\alias{tse}
\title{Example TreeSummarizedExperiment Data (TSE)}
\format{
A \code{TreeSummarizedExperiment} object.
}
\source{
Example dataset from microbiome analysis.
}
\usage{
data(tse)
}
\description{
A dataset containing microbiome data stored as a TreeSummarizedExperiment object.
}
\keyword{datasets}
