% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_differential_abundance.R
\name{visualize_differential_abundance}
\alias{visualize_differential_abundance}
\title{Visualize Differential Abundance Results}
\usage{
visualize_differential_abundance(
  final_results,
  group_var,
  point_size = 2,
  palette = MG()
)
}
\arguments{
\item{final_results}{The final results from the differential abundance analysis.}

\item{group_var}{The grouping variable used in the analysis.}

\item{point_size}{The size of points in the volcano plot. Default is 2.}

\item{palette}{A character vector specifying the color palette. Default is MG().}
}
\value{
A ggplot2 object representing the volcano plot.
}
\description{
Visualize Differential Abundance Results
}
